USF-4060 Bump @dropins/storefront-pdp to 3.2.0-beta.1 (#1245)
