+++ package.json
@@ -51,7 +51,7 @@
     "@dropins/storefront-checkout": "^3.2.1",
     "@dropins/storefront-order": "^3.3.0",
     "@dropins/storefront-payment-services": "^3.1.0",
-    "@dropins/storefront-pdp": "3.0.4-beta2",
+    "@dropins/storefront-pdp": "3.2.0-beta.1",
     "@dropins/storefront-personalization": "^3.1.1",
     "@dropins/storefront-product-discovery": "3.1.1-beta2",
     "@dropins/storefront-recommendations": "^4.0.3",
