+++ scripts/__dropins__/storefront-pdp/CHANGELOG.md
@@ -1,0 +1,27 @@
+# @dropins/storefront-pdp
+
+## 3.2.0-beta.1
+
+### Minor Changes
+
+- 763e5f9: Use the correct control when multi is true vs false so bundle options work without storefront customization.
+
+  - **multi === false:** Keep the existing single-select dropdown (`Picker`) for dropdown-type options; swatches unchanged for text/image/color.
+  - **multi === true:** Render Elsie `Checkbox` per option value (no radio or multi-select dropdown OOTB).
+  - **Swatches:** Selection state is either `{ label, value }` (single) or `{ label, values[] }` (multi); validation and labels updated accordingly.
+  - **Data / API:** `selectionMapToOptionUIDs` flattens selections and, when given `data.options`, orders UIDs by option groups; bundle completion uses “every required group has ≥1 child UID,” not `uids.length === option count`.
+  - **Containers:** `ProductOptions`, `ProductDetails`, and `initialize` use the shared helpers for payloads, validity, and placeholder filtering.
+  - **`getOptionUIDs` (bundles):** Prefer client `optionUIDs` when present; support multiple defaults when `multi` is true.
+  - **Tests:** Checkbox coverage for `multiple`, lib unit tests, and `ProductDetails` expectation aligned with stable UID order.
+
+- f59e497: Bump @adobe-commerce/elsie to 1.9.0-beta.0 and upgrade CI workflows to storefront-workflows v6 (Node 24)
+
+## 3.1.0
+
+### Minor Changes
+
+- 5c5be95: Add jest-preset-preact to devDependencies to fix test execution under Yarn Berry
+
+### Patch Changes
+
+- dc086a9: Resolve axios and flatted CVEs via yarn resolutions
