--- scripts/__dropins__/storefront-pdp/chunks/DownloadableOptions2.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{jsx as l,jsxs as s}from"@dropins/tools/preact-jsx-runtime.js";import{useState as $,useCallback as O}from"@dropins/tools/preact-compat.js";import{classes as j}from"@dropins/tools/lib.js";import{Price as T}from"@dropins/tools/components.js";/* empty css                   */import{useText as c}from"@dropins/tools/i18n.js";const B=({label:b,value:m,onLinkChange:t,currency:_,locale:w,className:u,...h})=>{const p=c("PDP.DownloadableOptions.Sample.label").label,N=c("PDP.DownloadableOptions.ViewSample.label"),{links:i,samples:f,purchaseSeparately:a}=m,S=a?new Set:new Set(i.map(e=>e.uid)),[v,D]=$(S),r=O((e,o,d)=>{!a||d||(D(P=>{const n=new Set(P);return o?n.add(e):n.delete(e),n}),t==null||t(e,o))},[t,a]),y=e=>{const o=a&&v.has(e.uid);return s("div",{className:`pdp-downloadable-options__item ${o?"pdp-downloadable-options__item--selected":""} ${a?"":"pdp-downloadable-options__item--disabled"}`,children:[l("button",{type:"button",className:"pdp-downloadable-options__clickable","aria-pressed":o,disabled:!a,onClick:()=>a&&r(e.uid,!o,!a),onKeyDown:d=>{a&&(d.key==="Enter"||d.key===" ")&&(d.preventDefault(),r(e.uid,!o,!a))},children:l("div",{className:"pdp-downloadable-options__content",children:s("div",{className:"pdp-downloadable-options__header",children:[l("span",{className:"pdp-downloadable-options__title",children:e.label}),l(T,{amount:e.price,currency:_,locale:w,className:"pdp-downloadable-options__price"})]})})}),e.sample_url&&l("a",{href:e.sample_url,className:"pdp-downloadable-options__sample-button",target:"_blank",rel:"noopener noreferrer","aria-label":N.label.replace("{label}",e.label),children:p})]},e.uid)},x=e=>s("div",{className:"pdp-downloadable-options__item pdp-downloadable-options__item--sample",children:[l("button",{type:"button",className:"pdp-downloadable-options__clickable","aria-pressed":!1,disabled:!0,children:l("div",{className:"pdp-downloadable-options__content",children:l("div",{className:"pdp-downloadable-options__header",children:l("span",{className:"pdp-downloadable-options__title",children:e.label})})})}),l("a",{href:e.url,className:"pdp-downloadable-options__sample-button",target:"_blank",rel:"noopener noreferrer","aria-label":`${p} ${e.label}`,children:p})]},e.uid);return l("div",{...h,className:j(["pdp-downloadable-options",u]),"data-slot-key":"product-downloadable-options",children:s("div",{className:"pdp-downloadable-options__field",children:[l("div",{className:"pdp-downloadable-options__field__label",children:b}),s("div",{className:"pdp-downloadable-options__list",children:[i.map(e=>y(e)),f.map(e=>x(e))]})]})})};export{B as D};
-//# sourceMappingURL=DownloadableOptions2.js.map
