--- scripts/__dropins__/storefront-pdp/chunks/GiftCardOptions2.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{jsxs as x,jsx as h}from"@dropins/tools/preact-jsx-runtime.js";import{useState as y,useCallback as M,useMemo as B,useEffect as q}from"@dropins/tools/preact-hooks.js";import{Field as V,TextArea as j,Picker as $,Input as z}from"@dropins/tools/components.js";import{classes as H}from"@dropins/tools/lib.js";/* empty css               */import{useText as u}from"@dropins/tools/i18n.js";const J=c=>{switch(c){case"number":return"number";case"email":return"email";default:return"text"}},K=c=>c.name==="giftcard/giftcard_message"?"textarea":c.name==="giftcard/giftcard_amount"?"picker":"input",Q=(c,a,i,t)=>[{text:i,value:"0",disabled:!1},...c.map(g=>({text:g.toString(),value:g.toString(),disabled:!1})),...a?[{text:t,value:"other",disabled:!1}]:[]],C=(c,a,i)=>{const{validationRule:t}=c;if(c.name==="giftcard/giftcard_amount"){if(t.required&&(a==="0"||!a||a.trim()==="")||a==="0")return i.RequiredFieldError;if(a==="other")return""}else if(t.required&&(!a||a.trim()===""))return i.RequiredFieldError;if(!a||a.trim()==="")return"";switch(t.type){case"email":{if(!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(a))return i.EmailError;break}case"number":{const l=Number(a);if(isNaN(l))return i.NumberError;if(t.min!==void 0&&l<t.min)return i.MinError.replace("{min}",t.min.toString());if(t.max!==void 0&&l>t.max)return i.MaxError.replace("{max}",t.max.toString());if(t.values&&!t.values.some(_=>typeof _=="number"?_===l:_===a))return i.InvalidValueError.replace("{values}",t.values.join(", "));break}case"string":{const l=a.length;if(t.minLength!==void 0&&l<t.minLength)return i.MinLengthError.replace("{min_length}",t.minLength.toString());if(t.maxLength!==void 0&&l>t.maxLength)return i.MaxLengthError.replace("{max_length}",t.maxLength.toString());break}}return""},tr=({options:c,currency:a,optionValues:i,onOptionChange:t,onValidationChange:l})=>{const g=u("PDP.Product.GiftCardOptions.ChooseAmount.label").label,_=u("PDP.Product.GiftCardOptions.OtherAmount.label").label,w=u("PDP.Product.GiftCardOptions.RequiredFieldError.label").label,N=u("PDP.Product.GiftCardOptions.NumberError.label").label,O=u("PDP.Product.GiftCardOptions.MinError.label").label,G=u("PDP.Product.GiftCardOptions.MaxError.label").label,L=u("PDP.Product.GiftCardOptions.MinLengthError.label").label,S=u("PDP.Product.GiftCardOptions.MaxLengthError.label").label,D=u("PDP.Product.GiftCardOptions.EmailError.label").label,R=u("PDP.Product.GiftCardOptions.InvalidValueError.label").label,[F,k]=y({}),[b,T]=y(!1),n=M(r=>c.find(e=>e.name===r),[c]),p=B(()=>({RequiredFieldError:w,NumberError:N,MinError:O,MaxError:G,MinLengthError:L,MaxLengthError:S,EmailError:D,InvalidValueError:R}),[w,N,O,G,L,S,D,R]);q(()=>{const r=c.every(e=>{if(e.name==="giftcard/custom_giftcard_amount"&&!b)return!0;const d=i.find(f=>f.id===e.uid),o=(d==null?void 0:d.value)||"";return C(e,o,p)===""});l(r)},[c,i,l,p,b]),q(()=>{const r=n("giftcard/giftcard_amount"),e=n("giftcard/custom_giftcard_amount"),d=i.find(E=>E.id===(r==null?void 0:r.uid)),o=(d==null?void 0:d.value)||"",f=b,P=r?o==="other":!!e;T(P),f&&!P&&e&&t(e.uid,"")},[i,n,b,t]);const v=(r,e)=>{const d=C(r,e,p);k(o=>({...o,[r.uid]:d})),t(r.uid,e)},A=(r,e)=>{const d=C(r,e,p);k(o=>({...o,[r.uid]:d}))},I=(r,e)=>{const d=e.target.value;v(r,d)},m=r=>{const e=i.find(s=>s.id===r.uid),d=e==null?void 0:e.value,o=F[r.uid],f=K(r),P=!!n("giftcard/custom_giftcard_amount");let E=r.label;return r.name==="giftcard/custom_giftcard_amount"&&a&&(E=`${r.label} ${a}`),h(V,{label:E,className:H(["pdp-gift-card-options__option-field",f==="textarea"&&"pdp-gift-card-options__option-field--full-width"]),error:o,children:f==="textarea"?h(j,{value:d,label:r.label,onChange:s=>v(r,s.target.value),onBlur:s=>A(r,s.target.value),"aria-invalid":!!o}):f==="picker"?h($,{"aria-label":r.label,label:r.label,handleSelect:s=>I(r,s),options:Q(r.validationRule.values||[],P,g,_),value:d,"aria-invalid":!!o}):h(z,{type:J(r.validationRule.type),value:d,onChange:s=>v(r,s.target.value),onBlur:s=>A(r,s.target.value),"aria-invalid":!!o})},r.uid)};return x("div",{className:"pdp-gift-card-options__grid",children:[x("div",{className:"pdp-gift-card-options__row",children:[n("giftcard/giftcard_amount")&&m(n("giftcard/giftcard_amount")),n("giftcard/custom_giftcard_amount")&&b&&m(n("giftcard/custom_giftcard_amount"))]}),x("div",{className:"pdp-gift-card-options__row",children:[n("giftcard/giftcard_sender_name")&&m(n("giftcard/giftcard_sender_name")),n("giftcard/giftcard_sender_email")&&m(n("giftcard/giftcard_sender_email"))]}),x("div",{className:"pdp-gift-card-options__row",children:[n("giftcard/giftcard_recipient_name")&&m(n("giftcard/giftcard_recipient_name")),n("giftcard/giftcard_recipient_email")&&m(n("giftcard/giftcard_recipient_email"))]}),h("div",{className:"pdp-gift-card-options__row",children:n("giftcard/giftcard_message")&&m(n("giftcard/giftcard_message"))})]})};export{tr as G};
-//# sourceMappingURL=GiftCardOptions2.js.map
