+++ scripts/__dropins__/storefront-pdp/chunks/components.js
@@ -1,0 +1,4 @@
+/*! Copyright 2026 Adobe
+All Rights Reserved. */
+import{debounce as xe,classes as A,VComponent as O}from"@dropins/tools/lib.js";import{useText as k}from"@dropins/tools/i18n.js";import*as he from"@dropins/tools/preact-compat.js";import{useState as V,Children as ce,useRef as re,useMemo as we,useEffect as U,useCallback as le,useLayoutEffect as qe}from"@dropins/tools/preact-compat.js";import{jsxs as $,jsx as t,Fragment as Pe}from"@dropins/tools/preact-jsx-runtime.js";import{Button as ge,Icon as Oe,Price as _e,Checkbox as Xe,Picker as ke,ColorSwatch as Ye,ImageSwatch as ze,TextSwatch as Fe,Field as He,TextArea as Ge,Input as Ie}from"@dropins/tools/components.js";import{cloneElement as je}from"@dropins/tools/preact.js";import{useState as Ae,useCallback as Ze,useMemo as Ke,useEffect as Se}from"@dropins/tools/preact-hooks.js";const Je=r=>he.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...r},he.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M7.74512 9.87701L12.0001 14.132L16.2551 9.87701",stroke:"currentColor",strokeWidth:1,strokeLinecap:"square",strokeLinejoin:"round"})),$e=r=>he.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...r},he.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M18.3599 5.64001L5.62988 18.37",stroke:"currentColor"}),he.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M18.3599 18.37L5.62988 5.64001",stroke:"currentColor"})),Qe=24,Le=16,Ve="32";var De=(r=>(r[r.small=16]="small",r[r.medium=24]="medium",r[r.large=64]="large",r))(De||{});const be=({show:r=1,scrollbar:s=!1,peak:i=!1,arrows:a=!1,controls:n="dots",arrowsOnMainImage:b=!1,loop:o=!1,gap:f=null,direction:S="horizontal",width:Y="100%",height:p="100%",defaultIndex:E=0,className:J,children:T,thumbnails:W=null,isZoomed:z,...j})=>{const Z=k("PDP.Carousel.label").label,y=k("PDP.Carousel.Slide.label").label,w=k("PDP.Carousel.Previous.label").label,e=k("PDP.Carousel.Next.label").label,h=k("PDP.Carousel.Controls.label").label,u=k("PDP.Carousel.Controls.Button.label").label,[m,x]=V(()=>Ee(r)),l=Math.ceil(ce.count(T)/m),_=i?Qe:0,g=f?De[f]:0,v=ce.count(T)%m,[N,q]=V(o?te()?E-1:E+1:E),F=re([]),[B,ie]=V(0),[ne,ue]=V(0),de=we(()=>ce.toArray(W),[W]),c=we(()=>{const d=ce.toArray(T).reduce((P,R,L)=>L%m===0?[...P,[R]]:[...P.slice(0,-1),[...P[P.length-1],R]],[]);return!o||d.length===1?d:o&&ce.count(T)%m===0?[d[l-1]].concat(d).concat([d[0],d[1]]):[ce.toArray(T).slice(-m)].concat(d).concat([d[0],d[1]])},[T,m,l,o]),I=re(null),X=re(null),D=n==="thumbnailsRow"||n==="thumbnailsColumn";U(()=>{const d=I.current;if(!d)return;const P=xe(()=>{ie(d.offsetWidth??0),ue(d.offsetHeight??0);const L=window.innerWidth,G=Ee(r,L);G!==m&&x(G)},50),R=new ResizeObserver(P);return R.observe(d),()=>{R.unobserve(d)}},[m,r]);const H=le(d=>{const P=o?d-1:d,R=F==null?void 0:F.current[0],L=(R==null?void 0:R.offsetWidth)+Le,G=(R==null?void 0:R.offsetHeight)+Le,K=X.current,pe=n==="thumbnailsRow",me=n==="thumbnailsColumn",se=pe?K==null?void 0:K.offsetWidth:K==null?void 0:K.offsetHeight,ee=pe?L:G;if((pe||me)&&se){const Te=Math.floor((pe?K.clientWidth:K.clientHeight)/ee),We=Math.floor(Te/2),fe=(P-We)*ee;K.style.scrollBehavior="smooth",pe?K.scrollLeft=te()?-(K.scrollWidth-K.clientWidth-fe):fe:K.scrollTop=fe}},[n,o]);U(()=>{H(N)},[N,H]);const M=le((d,P,R)=>{D&&H(d);const L=I.current,G=_?L.offsetWidth-_:L.offsetWidth-_+g;if(!Array.from(L.querySelectorAll("[data-index]")).filter(se=>{const ee=Number(se.getAttribute("data-index"));return!isNaN(ee)&&ee>-1})[d])return;if(R){const se=d===0?l:0,ee=G*(l+(v||m)/m);if(te()){if(L.scrollLeft<=0&&L.scrollLeft>-(G-5)){M(1,!0);return}if(L.scrollLeft<G-ee&&L.scrollLeft>=-ee){M(l,!0);return}L.style.scrollBehavior="auto",L.scrollLeft=se===0?0:-ee,M(se===0?1:l,!0)}else{if(L.scrollLeft>=0&&L.scrollLeft<G-5){M(1,!0);return}if(L.scrollLeft>ee-G&&L.scrollLeft<=ee){M(l,!0);return}L.style.scrollBehavior="auto",L.scrollLeft=se===0?0:ee,M(se===0?1:l,!0);return}}const me=(L.offsetWidth-g)/m*c[d].length-_;L.style.scrollBehavior=P?"smooth":"auto",L.scrollLeft=te()?-(c[d].length===m?G*d:G*(d-1)-me):c[d].length===m?G*d:G*(d-1)+me},[g,l,_,m,v,c,H,n]),Q=le(()=>{if(o){const P=N-1;M(P,!0,te()?N===l-1:N===1);return}const d=N<=0?0:N-1;M(d,!0)},[N,M,o]),C=le(()=>{const d=c.length-1;if(o){const R=N+1;M(R,!0,te()?N===0:N===l);return}const P=N>=d?d:N+1;M(P,!0)},[c.length,o,N,M,l]);U(()=>{const d=P=>{P.key==="ArrowLeft"&&(te()?C():Q()),P.key==="ArrowRight"&&(te()?Q():C())};return window.addEventListener("keydown",d),()=>{window.removeEventListener("keydown",d)}},[C,Q]),qe(()=>{M(o?E+1:E,!1)},[E,M,o,B]);const oe=xe(()=>{const d=I.current,P=_?d.offsetWidth-_:d.offsetWidth-_+g,R=Math.abs(d.scrollLeft)/P,L=Math.abs(Math.round(R)-R<1?Math.round(R):Math.ceil(R));if(D&&H(L),o){const G=P*(l+(v||m)/m),K=te()?-d.scrollLeft:d.scrollLeft;if(Math.ceil(Math.abs(K))>=Math.ceil(G)-5){d.style.scrollBehavior="auto",d.scrollLeft=te()?-d.offsetWidth+(Math.ceil(K)-Math.ceil(G)-_):d.offsetWidth+(Math.ceil(K)-Math.ceil(G)-_);return}if(Math.abs(K)===0){d.style.scrollBehavior="auto",d.scrollLeft=te()?-(G-d.offsetWidth+_):G-d.offsetWidth+_;return}}N!==L&&q(L)},100);U(()=>{const d=I.current;return d==null||d.addEventListener("scroll",oe),()=>{d==null||d.removeEventListener("scroll",oe)}},[oe]);const ae=(d,P,R)=>t(ge,{className:A([["pdp-carousel__button",!d],[`pdp-carousel__button--${P}`,!d],[`pdp-carousel__button--${n}`,d]]),style:{"--height":`${ne}px`},variant:"tertiary","aria-label":P==="next"?e:w,onClick:R,disabled:o?!1:P==="next"?N>=c.length-1:N<1,children:t(Oe,{className:A([["pdp-carousel__button__icon",!d],[`pdp-carousel__button__icon--${P}`,!d],[`pdp-carousel__button__icon--${n}--${P}`,d]]),size:Ve,source:Je})}),Ne=a&&c.length!=1&&ae(D,"prev",Q),ye=a&&c.length!=1&&ae(D,"next",C),Re=a&&c.length!=1&&ae(!1,"prev",Q),Be=a&&c.length!=1&&ae(!1,"next",C);return $("div",{role:"region","aria-roledescription":Z,className:A(["pdp-carousel",["pdp-carousel--main-image-controls",b],["pdp-carousel--arrows",a&&!D],[`pdp-carousel--${n}`,D],J]),style:{"--flex-carousel":n==="thumbnailsColumn"?"row-reverse":"column","--gap":D?"0":"var(--spacing-small)","--width":Y},...j,children:[t("div",{ref:I,className:A(["pdp-carousel__wrapper",`pdp-carousel__wrapper--${S}`,["pdp-carousel__wrapper--scrollbar",s],["pdp-carousel__wrapper--peak",i]]),style:{"--total-width":n==="thumbnailsColumn"?"81.6%":"100%","--height":n==="thumbnailsColumn"?"auto":p,"--gap":f?`var(--spacing-${f})`:"0px","--per-page":m},tabIndex:0,children:c.map((d,P)=>{const R=d.length<m;return t("div",{role:"group","aria-roledescription":y,"data-index":P,className:A(["pdp-carousel__slide",`pdp-carousel__slide--${S}`,["pdp-carousel__slide--active",N===P],["pdp-carousel__slide--orphan",R],["pdp-carousel__slide--hidden",B===0]]),style:{"--length":d.length},children:d},P)})}),!D&&$(Pe,{children:[Ne,ye]}),D&&b&&$(Pe,{children:[Re,Be]}),n&&c.length!=1&&$("div",{className:A([[`pdp-carousel__controls__container--${n}`,D],["pdp-carousel__controls__container--no-arrows",!a||b]]),style:{"--width":n==="thumbnailsRow"?`${B}px`:Y,"--height":`${ne}px`,"--nr-thumbnails":de.length},children:[(D||z)&&!b&&Ne,t("div",{ref:X,className:A([[`pdp-carousel__controls__wrapper--${n}`,D],["pdp-carousel__controls__wrapper",D]]),children:t("div",{role:"group","aria-label":h,className:A(["pdp-carousel__controls",[`pdp-carousel__controls--${n}`,D]]),children:(o?c.slice(0,l):c).map((d,P)=>{const R=o?l:c.length,L=o?N-1===P:N===P;return D?$("label",{className:A(["pdp-carousel__thumbnail__container"]),ref:G=>{G&&(F.current[P]=G)},children:[t("input",{type:"radio",name:"carousel-thumbnails","aria-label":u.replace("{key}",String(P+1)).replace("{total}",String(R)),checked:L&&B!==0,onChange:()=>{M(o?P+1:P,!0)},...j,className:A(["pdp-carousel__thumbnail",["pdp-carousel__thumbnail--selected",L&&B!==0]])}),t("span",{className:A(["pdp-carousel__thumbnail__span"]),children:de[P]})]},P):t("button",{"aria-label":u.replace("{key}",String(P+1)).replace("{total}",String(R)),onClick:()=>{M(o?P+1:P,!0)},className:A(["pdp-carousel__controls__button",["pdp-carousel__controls__button--active",L]])},P)})})}),(D||z)&&!b&&ye]})]})};function Ee(r,s=0){if(typeof r=="number")return r;const{small:i,medium:a,large:n}=r,o=[{breakpoint:1024,value:n},{breakpoint:786,value:a},{breakpoint:0,value:i}].find(({breakpoint:f})=>s>f);return o?o.value:i}function te(){return document.documentElement.dir==="rtl"}const _t=({title:r,breadcrumbs:s,galleryContent:i,infoContent:a,productContent:n,sku:b,outOfStock:o,shortDescription:f,description:S,attributes:Y,images:p,options:E,giftCardOptions:J,downloadableOptions:T,price:W,specialPrice:z,className:j,children:Z,quantity:y,actions:w,carouselConfig:e,zoomType:h="overlay",closeButton:u=!0,...m})=>{var B,ie,ne,ue,de,c,I,X,D;const x=k("PDP.Product.RegularPrice.label").label,l=k("PDP.Product.SpecialPrice.label").label,_=k("PDP.Product.OutOfStock.label").label,g=k("PDP.Product.Image.label").label,[v,N]=V(null),q=p!=null&&p.length?p.map((H,M)=>{var C;const Q=((C=r==null?void 0:r.props)==null?void 0:C.children)&&(g==null?void 0:g.replace("{product}",r.props.children.toString()).replace("{key}",String(M+1)).replace("{total}",String(p.length)));return h==="zoom"?t(lt,{closeButton:u,children:t(O,{node:H,loading:M===0?"eager":"lazy",alt:Q},M)}):t(O,{node:H,loading:M===0?"eager":"lazy",alt:Q,onClick:()=>N(M)},M)}):t("img",{src:"data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7",className:"pdp-product__images__placeholder","aria-hidden":!0}),F=(B=e==null?void 0:e.thumbnails)!=null&&B.length?e.thumbnails.map((H,M)=>{var Q;return t(O,{node:H,loading:(e==null?void 0:e.thumbnailsLoadingMode)||"lazy",alt:((Q=r==null?void 0:r.props)==null?void 0:Q.children)&&(g==null?void 0:g.replace("{product}",r.props.children.toString()).replace("{key}",String(M+1)).replace("{total}",String(p.length)))},M)}):t("img",{src:"data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7",className:"pdp-product__images__placeholder","aria-hidden":!0});return $("div",{...m,className:A(["pdp-product",j]),children:[s&&t(O,{node:s,className:"pdp-product__breadcrumbs"}),$("div",{className:"pdp-product__column-container pdp-product__gallery-column",children:[t("div",{className:"pdp-product__column-body",children:(ie=e==null?void 0:e.controls)!=null&&ie.desktop?t(be,{className:"pdp-product__overlay__carousel",arrows:(p==null?void 0:p.length)>1,controls:(p==null?void 0:p.length)>1?e.controls.desktop:"dots",arrowsOnMainImage:e==null?void 0:e.arrowsOnMainImage,loop:(e==null?void 0:e.loopable)&&(p==null?void 0:p.length)>1,peak:((ne=e==null?void 0:e.thumbnails)==null?void 0:ne.length)>1&&((ue=e==null?void 0:e.peak)==null?void 0:ue.desktop)||!1,gap:(e==null?void 0:e.gap)||null,width:"100%",height:"100%",defaultIndex:v||0,thumbnails:F,children:q}):t(Ue,{className:"pdp-product__images",gap:"small",children:q})}),i&&t(O,{node:i,className:"pdp-product__gallery-content"})]}),$("div",{className:"pdp-product__column-container pdp-product__content-column",children:[$("div",{className:"pdp-product__column-body",children:[$("div",{className:"pdp-product__header",children:[r&&t(O,{node:r,className:"pdp-product__title"}),b&&t(O,{node:b,className:"pdp-product__sku"})]}),$("div",{className:"pdp-product__prices",children:[z&&t(O,{node:z,className:A(["pdp-product__price-special","pdp-product__price"]),"aria-label":l,role:"text"}),W&&t(O,{node:W,className:A(["pdp-product__price-regular","pdp-product__price--grey"]),"aria-label":x,role:"text"})]}),(de=e==null?void 0:e.controls)!=null&&de.mobile&&((c=e==null?void 0:e.controls)==null?void 0:c.mobile)==="thumbnailsRow"?t(be,{className:"pdp-product__images pdp-product__images--carousel pdp-product__images--carousel--thumbnails",width:"100%",height:"auto",arrows:(p==null?void 0:p.length)>1,controls:(p==null?void 0:p.length)>1?"thumbnailsRow":null,loop:(e==null?void 0:e.loopable)&&(p==null?void 0:p.length)>1,peak:((I=e==null?void 0:e.thumbnails)==null?void 0:I.length)>1&&((X=e==null?void 0:e.peak)==null?void 0:X.mobile)||!1,gap:(e==null?void 0:e.gap)||null,thumbnails:F,children:q}):t(be,{className:"pdp-product__images pdp-product__images--carousel",peak:((D=e==null?void 0:e.peak)==null?void 0:D.mobile)&&(p==null?void 0:p.length)>1,gap:(e==null?void 0:e.gap)||null,width:"100%",height:"auto",arrows:(p==null?void 0:p.length)>1,controls:(p==null?void 0:p.length)>1?"dots":null,loop:(e==null?void 0:e.loopable)&&(p==null?void 0:p.length)>1,children:q}),$("div",{className:A(["pdp-product__actions",["pdp-product__actions--out-of-stock",o]]),children:[o&&t("div",{className:"pdp-product__out-of-stock__text",children:_}),E&&t(O,{node:E,className:A(["pdp-product__options"])}),J&&t(O,{node:J,className:A(["pdp-product__gift-card-options"])}),T&&t(O,{node:T,className:A(["pdp-product__downloadable-options"])}),y&&t("div",{className:"pdp-product__quantity",children:t(O,{node:y})}),w&&t(O,{node:w,className:A(["pdp-product__buttons"])})]}),f&&t(O,{node:f,className:A(["pdp-product__short_description"])}),S&&t(O,{node:S,className:A(["pdp-product__description"])}),Y?t("div",{className:"pdp-product__attributes",children:Y}):null]}),a&&t(O,{node:a,className:"pdp-product__info-content"})]}),n&&t(O,{node:n,className:"pdp-product__content"}),v===null?null:t(Ce,{className:"pdp-product__overlay",onClose:()=>N(null),centered:!0,children:t(be,{className:"pdp-product__overlay__carousel",arrows:(p==null?void 0:p.length)>1,width:"100%",height:"100%",defaultIndex:v||0,loop:(e==null?void 0:e.loopable)&&(p==null?void 0:p.length)>1,controls:(p==null?void 0:p.length)>1?"dots":null,isZoomed:!0,children:q})})]})},Ue=({children:r,gap:s=null,className:i,style:a,...n})=>{const b=ce.count(r);return t("div",{...n,className:A(["pdp-gallery-grid",i]),style:{"--gap":s?`var(--spacing-${s})`:"0px",...a},children:ce.map(r,(o,f)=>t("div",{className:"pdp-gallery-grid__item",children:t(O,{node:o,loading:f===0?"eager":"lazy"},o.props.src)},b+f))})},mt=({amount:r,currency:s,locale:i,variant:a,sale:n,minimumAmount:b,maximumAmount:o,className:f,...S})=>t(Pe,{children:r||b===o?t("div",{className:"pdp-price-range",...S,children:t(_e,{amount:r||b,currency:s,locale:i,variant:a,sale:n,className:f})}):$("div",{className:"pdp-price-range",...S,children:[t(_e,{amount:b,currency:s,locale:i,className:f}),t("span",{className:"pdp-price-range__label",children:"-"}),t(_e,{amount:o,currency:s,locale:i,className:f})]})}),Ce=({centered:r,onClose:s,className:i,children:a,...n})=>{const b=k("PDP.Overlay.Close.label").label,o=re(null),f=le(()=>{s==null||s()},[s]);return U(()=>{const S=Y=>{Y.key==="Escape"&&f()};return document.addEventListener("keydown",S),()=>{document.removeEventListener("keydown",S)}},[f]),U(()=>{const S=o.current;S==null||S.focus();const Y=document.scrollingElement,p=Y.style.overflow;return Y.style.overflow="hidden",()=>{Y.style.overflow=p}},[]),$("div",{...n,className:A(["pdp-overlay",["pdp-overlay--centered",r],i]),ref:o,role:"dialog","aria-describedby":"overlay-content",tabindex:0,children:[t("div",{className:"pdp-overlay__content",id:"overlay-content",children:a}),t(ge,{"aria-label":b,variant:"tertiary",className:"pdp-overlay__close-button",onClick:f,icon:t($e,{})})]})},et=["text","image","color","dropdown"],tt=r=>r.map(s=>({...s,text:s.label,disabled:!s.inStock}));function Me(r,s){return s.reduce((i,a)=>{const n=a.items.filter(o=>r.includes(o.id));if(n.length===0)return i;if(a.multiple)return{...i,[a.id]:{label:n.map(o=>o.label).join(", "),values:n.map(o=>o.id)}};const b=n[0];return{...i,[a.id]:{label:b.label,value:b.id}}},{})}function rt(r){if(!(!r||!("value"in r)))return r.value}const bt=({options:r,hideSelectedValue:s,disablePreselections:i,selectedUIDs:a,onValues:n,onErrors:b,defaultOptions:o,selectionsToUpdate:f,className:S,children:Y,imageSwatchNode:p,...E})=>{const J=k("PDP.Swatches.Required.label").label,T=k("PDP.Swatches.ChooseOption.label").label,[W,z]=V(()=>{let e=o!=null&&o.length&&(r!=null&&r.length)?Me(o,r):{};return r==null||r.forEach(h=>{if(e[h.id])return;const{items:u}=h;if(h.multiple){const l=u==null?void 0:u.filter(_=>!!(o!=null&&o.includes(_.id))||_.selected);l!=null&&l.length&&(e={...e,[h.id]:{label:l.map(_=>_.label).join(", "),values:l.map(_=>_.id)}});return}const m=u==null?void 0:u.find(l=>o==null?void 0:o.includes(l.id)),x=u==null?void 0:u.find(l=>l.selected);m?e={...e,[h.id]:{label:m.label,value:m.id}}:x&&(e={...e,[h.id]:{label:x.label,value:x.id}})}),e});U(()=>{if(a){const e=Me(a,r);if(JSON.stringify(e)===JSON.stringify(W))return;z(h=>{let u={...h,...e};return n==null||n(u,(a==null?void 0:a[0])??""),u})}},[a]),U(()=>{f==null||f.forEach(e=>{z(h=>{var l;const u=(l=e==null?void 0:e.items)==null?void 0:l.filter(_=>_.selected);if(!(u!=null&&u.length))return h;if(e.multiple)return{...h,[e.id]:{label:u.map(g=>g.label).join(", "),values:u.map(g=>g.id)}};const m=u[0];return{...h,[e.id]:{label:m.label,value:m.id}}})})},[f,z]),we(()=>{const e=r==null?void 0:r.reduce((h,u)=>{var l;const m=W[u.id];let x=!1;return u.multiple?x=!!(m&&"values"in m&&m.values.length>0):(l=u==null?void 0:u.items)==null||l.forEach(_=>{m&&"value"in m&&m.value===_.id&&(x=!0)}),u.required&&!x&&(h[u.id]=J),h},{});return b==null||b(e),e},[W]);const j=le((e,h,u)=>{z(m=>{let x={...m,...e&&{[e]:{label:h,value:u}}};if(h===T){const l={...x};delete l[e],x=l}return n==null||n(x,u),x})},[n]),Z=le((e,h,u,m)=>{z(x=>{const l=x[e],_=l&&"values"in l?[...l.values]:l&&"value"in l&&l.value?[l.value]:[],g=u?[...new Set([..._,h])]:_.filter(F=>F!==h),v=new Map(m.map(F=>[F.id,F.label])),N=g.map(F=>v.get(F)).filter(Boolean).join(", "),q={...x};return g.length===0?delete q[e]:q[e]={label:N,values:g},n==null||n(q,h),q})},[n]),y=le((e,h)=>{var m;const u=((m=h.selectedOptions[0])==null?void 0:m.label)??"";j(e,u,h.value)},[j]),w=le(({items:e,id:h,required:u,type:m,selected:x,label:l,multiple:_})=>{const g=`pdp-swatch-${h}=${Math.random().toString(36)}`;if(_){const v=W[h],N=new Set(v&&"values"in v?v.values:[]);return(e==null?void 0:e.length)>0&&t("div",{className:A(["pdp-swatches__options","pdp-swatches__options--checkboxes"]),children:e==null?void 0:e.map(q=>t(Xe,{name:`pdp-bundle-multi__${h}__${q.id}`,label:q.label,checked:N.has(q.id),disabled:!q.inStock,onChange:F=>Z(h,q.id,F.currentTarget.checked,e)},q.id))})}return(e==null?void 0:e.length)>0&&t("div",{className:A(["pdp-swatches__options"]),children:(m==null?void 0:m.toLowerCase())==="dropdown"?t(ke,{id:g,name:h,"aria-label":l,placeholder:i?T:void 0,handleSelect:v=>y(h,v.target),options:tt(e),value:x}):e==null?void 0:e.map(v=>{const N={key:v.id,id:`${g}-${v.id}`,name:h,value:v.value,label:v.label,groupAriaLabel:l,selected:x===v.id,outOfStock:!v.inStock,required:u===null?!0:u,onValue:()=>{j(h,v.label,v.id)}};switch(m){case"text":return t(Fe,{...N,label:v.label});case"image":return t(ze,{...N,src:v.value,alt:v.label,imageNode:p});case"color":return t(Ye,{...N,size:"large",color:v.value});default:return console.warn(`Invalid swatch type ${m}`),null}})})},[T,j,y,Z,W]);return t("div",{...E,className:A(["pdp-swatches",S]),children:r==null?void 0:r.map(({type:e,label:h,id:u,multiple:m,required:x,items:l})=>(et.includes(e)||console.warn(`Invalid swatch type ${e}`),$("div",{id:`swatch-item-${u}`,"data-slot-key":`product-swatch--${u}`,className:"pdp-swatches__field",children:[t("div",{className:"pdp-swatches__field__label",children:W[u]?s?`${h}`:`${h}: ${W[u].label}`:h}),w({id:u,type:e,required:x,items:l,label:h,multiple:m,selected:rt(W[u])})]},`swatch-item-${u}`)))})},lt=({children:r,scale:s=2,closeButton:i=!0})=>{var de;const[a,n]=V(!1),[b,o]=V("pdp-zoom"),[f,S]=V(!1),[Y,p]=V({x:0,y:0}),[E,J]=V({x:0,y:0}),[T,W]=V({x:0,y:0}),z=re(null),j=re(null),Z=re(null),y=re("none"),w=re(null),e=re(!1),h=re(null),u=le(c=>{const I=()=>{var ae;let X;j.current&&z.current?X=j.current:z.current?(X=z.current.getBoundingClientRect(),j.current=X):X=(ae=c.currentTarget)==null?void 0:ae.getBoundingClientRect();const D=(100-100/s)/2;let H=null;if("touches"in c)if(c.touches.length>0)H=c.touches[0],h.current={clientX:H.clientX,clientY:H.clientY};else if(c.type==="touchend"&&w.current)H={clientX:w.current.x,clientY:w.current.y};else if(h.current)H=h.current;else return;else if(c instanceof MouseEvent)H=c;else return;const M=(H.clientX-X.left)/X.width*100,Q=(H.clientY-X.top)/X.height*100,C=(50-M)*(s-1)/s,oe=(50-Q)*(s-1)/s;W({x:Math.max(-D,Math.min(C,D))||0,y:Math.max(-D,Math.min(oe,D))||0})};Z.current&&cancelAnimationFrame(Z.current),Z.current=requestAnimationFrame(I)},[s]);U(()=>{j.current=null},[a]);const m=c=>{!a||c.type!=="touchstart"||(p({x:c.touches[0].clientX,y:c.touches[0].clientY}),J({x:T.x,y:T.y}),S(!0))},x=c=>{var ae;if(!f||c.type!=="touchmove")return;const I=(ae=c.currentTarget)==null?void 0:ae.getBoundingClientRect(),X=(100-100/s)/2,D=c.touches[0].clientX-Y.x,H=c.touches[0].clientY-Y.y,M=D*100/(I.width*s),Q=H*100/(I.height*s),C=E.x+M,oe=E.y+Q;W({x:Math.max(-X,Math.min(C,X))||0,y:Math.max(-X,Math.min(oe,X))||0}),c.preventDefault()},l=c=>{if(i)w.current={x:c.touches[0].clientX,y:c.touches[0].clientY,time:Date.now()},h.current={clientX:c.touches[0].clientX,clientY:c.touches[0].clientY},e.current=!1,a&&m(c);else{if(f)return;!a&&u(c),n(!a),c.preventDefault()}},_=c=>{if(c.touches.length!==0){if(h.current={clientX:c.touches[0].clientX,clientY:c.touches[0].clientY},i){if(!w.current)return;if(f){x(c);return}const I=c.touches[0].clientX,X=c.touches[0].clientY,D=Math.abs(I-w.current.x),H=Math.abs(X-w.current.y);(D>10||H>10)&&(e.current=!0)}a&&u(c)}},g=c=>{if(i){if(!w.current)return;const I=Date.now()-w.current.time;!e.current&&I<300&&!f&&(n(!0),u(c),c.preventDefault()),S(!1),w.current=null}else n(!1)},v=c=>{"touches"in c||f||(a?!i&&n(!1):(n(!0),u(c)),c.preventDefault())},N=c=>{"touches"in c||a&&u(c)},q=c=>{if(!("touches"in c)){if(i){S(!1);return}n(!1)}},F=c=>{n(!1),c.stopPropagation()},B=a?`scale(${s}) translateX(${T.x}%) translateY(${T.y}%)`:"none",ie=k("PDP.Zoom.Close.label").label,ne=ce.only(r);U(()=>{const c=()=>{a&&(n(!1),W({x:0,y:0}),S(!1))},I=document.querySelector("#gallery");if(I)return I.addEventListener("click",c),()=>I.removeEventListener("click",c)},[a]);const ue=je(ne,{...ne.props,draggable:!1,style:{...(de=ne.props)==null?void 0:de.style,"--transform":B}});return U(()=>{o(a?`pdp-zoom ${i?"pdp-zoom--no-effects":"pdp-zoom--zoomed"}`:"pdp-zoom")},[a,i]),U(()=>{if(y.current===B)return;y.current=B,requestAnimationFrame(()=>{if(!z.current)return;z.current.querySelectorAll("img").forEach(X=>{a&&B!=="none"?X.style.setProperty("transform",B,"important"):X.style.removeProperty("transform")})})},[B,a]),U(()=>()=>{Z.current&&cancelAnimationFrame(Z.current)},[]),$("figure",{ref:z,className:b,onClick:v,onMouseMove:N,onMouseLeave:q,onTouchStart:l,onTouchMove:_,onTouchEnd:g,children:[ue,i&&a&&t(ge,{"aria-label":ie,variant:"tertiary",className:"pdp-zoom__close-button",icon:t($e,{}),onClick:F})]})},ft=({className:r,title:s,sku:i,hideSku:a=!1,variantName:n,variantSku:b,children:o,...f})=>$("div",{...f,className:A(["pdp-header",r]),children:[t("div",{className:"pdp-header__title",children:t(O,{node:s})}),!a&&t("div",{className:"pdp-header__sku",children:t(O,{node:i})}),(n||b)&&$("div",{className:"pdp-header__variant",children:[n&&t("span",{className:"pdp-header__variant-name",children:t(O,{node:n})}),n&&b&&t("span",{className:"pdp-header__variant-separator",children:" – "}),b&&t("span",{className:"pdp-header__variant-sku",children:t(O,{node:b})})]})]}),vt=({className:r,price:s,specialPrice:i,tiers:a,children:n,...b})=>{const o=k("PDP.Product.RegularPrice.label").label,f=k("PDP.Product.SpecialPrice.label").label,S=k("PDP.Product.TierPrice.label").label;return $("div",{...b,className:A(["pdp-price",r]),children:[i&&t(O,{node:i,className:A(["pdp-price__amount-special","pdp-price__amount"]),"aria-label":f,role:"text"}),s&&t(O,{node:s,className:A(["pdp-price__amount-regular","pdp-price__amount--grey"]),"aria-label":o,role:"text"}),a&&t(O,{node:a,className:A(["pdp-price__tiers"]),"aria-label":S,role:"text"})]})},wt=({locale:r,tiers:s,regularPrice:i,...a})=>{const n=k("PDP.Product.TierPricing.label").label;return!s||s.length===0?null:t("div",{className:"pdp-price-tiers",...a,children:s.map((b,o)=>{var T,W;const f=((T=b==null?void 0:b.tier)==null?void 0:T.amount)??0,S=(b==null?void 0:b.quantity)??0,Y=((W=b==null?void 0:b.tier)==null?void 0:W.currency)??"";let p=0;i&&f&&f<i&&(p=Math.round((i-f)/i*100));const E=n.replace("{quantity}",S.toString()).split("{price}")[0],J=n.replace("{quantity}",S.toString()).split("{price}")[1].replace("{percent}",p>0?p.toString():"0");return t("div",{className:"pdp-price-tiers__tier",children:$("span",{className:"pdp-price-tiers__text",children:[E,t(_e,{amount:f,currency:Y,locale:r,className:"pdp-price-tiers__price"}),J]})},o)})})},at=r=>{switch(r){case"number":return"number";case"email":return"email";default:return"text"}},nt=r=>r.name==="giftcard/giftcard_message"?"textarea":r.name==="giftcard/giftcard_amount"?"picker":"input",ct=(r,s,i,a)=>[{text:i,value:"0",disabled:!1},...r.map(b=>({text:b.toString(),value:b.toString(),disabled:!1})),...s?[{text:a,value:"other",disabled:!1}]:[]],ve=(r,s,i)=>{const{validationRule:a}=r;if(r.name==="giftcard/giftcard_amount"){if(a.required&&(s==="0"||!s||s.trim()==="")||s==="0")return i.RequiredFieldError;if(s==="other")return""}else if(a.required&&(!s||s.trim()===""))return i.RequiredFieldError;if(!s||s.trim()==="")return"";switch(a.type){case"email":{if(!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(s))return i.EmailError;break}case"number":{const n=Number(s);if(isNaN(n))return i.NumberError;if(a.min!==void 0&&n<a.min)return i.MinError.replace("{min}",a.min.toString());if(a.max!==void 0&&n>a.max)return i.MaxError.replace("{max}",a.max.toString());if(a.values&&!a.values.some(o=>typeof o=="number"?o===n:o===s))return i.InvalidValueError.replace("{values}",a.values.join(", "));break}case"string":{const n=s.length;if(a.minLength!==void 0&&n<a.minLength)return i.MinLengthError.replace("{min_length}",a.minLength.toString());if(a.maxLength!==void 0&&n>a.maxLength)return i.MaxLengthError.replace("{max_length}",a.maxLength.toString());break}}return""},Pt=({options:r,currency:s,optionValues:i,onOptionChange:a,onValidationChange:n})=>{const b=k("PDP.Product.GiftCardOptions.ChooseAmount.label").label,o=k("PDP.Product.GiftCardOptions.OtherAmount.label").label,f=k("PDP.Product.GiftCardOptions.RequiredFieldError.label").label,S=k("PDP.Product.GiftCardOptions.NumberError.label").label,Y=k("PDP.Product.GiftCardOptions.MinError.label").label,p=k("PDP.Product.GiftCardOptions.MaxError.label").label,E=k("PDP.Product.GiftCardOptions.MinLengthError.label").label,J=k("PDP.Product.GiftCardOptions.MaxLengthError.label").label,T=k("PDP.Product.GiftCardOptions.EmailError.label").label,W=k("PDP.Product.GiftCardOptions.InvalidValueError.label").label,[z,j]=Ae({}),[Z,y]=Ae(!1),w=Ze(l=>r.find(_=>_.name===l),[r]),e=Ke(()=>({RequiredFieldError:f,NumberError:S,MinError:Y,MaxError:p,MinLengthError:E,MaxLengthError:J,EmailError:T,InvalidValueError:W}),[f,S,Y,p,E,J,T,W]);Se(()=>{const l=r.every(_=>{if(_.name==="giftcard/custom_giftcard_amount"&&!Z)return!0;const g=i.find(N=>N.id===_.uid),v=(g==null?void 0:g.value)||"";return ve(_,v,e)===""});n(l)},[r,i,n,e,Z]),Se(()=>{const l=w("giftcard/giftcard_amount"),_=w("giftcard/custom_giftcard_amount"),g=i.find(F=>F.id===(l==null?void 0:l.uid)),v=(g==null?void 0:g.value)||"",N=Z,q=l?v==="other":!!_;y(q),N&&!q&&_&&a(_.uid,"")},[i,w,Z,a]);const h=(l,_)=>{const g=ve(l,_,e);j(v=>({...v,[l.uid]:g})),a(l.uid,_)},u=(l,_)=>{const g=ve(l,_,e);j(v=>({...v,[l.uid]:g}))},m=(l,_)=>{const g=_.target.value;h(l,g)},x=l=>{const _=i.find(B=>B.id===l.uid),g=_==null?void 0:_.value,v=z[l.uid],N=nt(l),q=!!w("giftcard/custom_giftcard_amount");let F=l.label;return l.name==="giftcard/custom_giftcard_amount"&&s&&(F=`${l.label} ${s}`),t(He,{label:F,className:A(["pdp-gift-card-options__option-field",N==="textarea"&&"pdp-gift-card-options__option-field--full-width"]),error:v,children:N==="textarea"?t(Ge,{value:g,label:l.label,onChange:B=>h(l,B.target.value),onBlur:B=>u(l,B.target.value),"aria-invalid":!!v}):N==="picker"?t(ke,{"aria-label":l.label,label:l.label,handleSelect:B=>m(l,B),options:ct(l.validationRule.values||[],q,b,o),value:g,"aria-invalid":!!v}):t(Ie,{type:at(l.validationRule.type),value:g,onChange:B=>h(l,B.target.value),onBlur:B=>u(l,B.target.value),"aria-invalid":!!v})},l.uid)};return $("div",{className:"pdp-gift-card-options__grid",children:[$("div",{className:"pdp-gift-card-options__row",children:[w("giftcard/giftcard_amount")&&x(w("giftcard/giftcard_amount")),w("giftcard/custom_giftcard_amount")&&Z&&x(w("giftcard/custom_giftcard_amount"))]}),$("div",{className:"pdp-gift-card-options__row",children:[w("giftcard/giftcard_sender_name")&&x(w("giftcard/giftcard_sender_name")),w("giftcard/giftcard_sender_email")&&x(w("giftcard/giftcard_sender_email"))]}),$("div",{className:"pdp-gift-card-options__row",children:[w("giftcard/giftcard_recipient_name")&&x(w("giftcard/giftcard_recipient_name")),w("giftcard/giftcard_recipient_email")&&x(w("giftcard/giftcard_recipient_email"))]}),t("div",{className:"pdp-gift-card-options__row",children:w("giftcard/giftcard_message")&&x(w("giftcard/giftcard_message"))})]})},gt=({label:r,value:s,onLinkChange:i,currency:a,locale:n,className:b,...o})=>{const f=k("PDP.DownloadableOptions.Sample.label").label,S=k("PDP.DownloadableOptions.ViewSample.label"),{links:Y,samples:p,purchaseSeparately:E}=s,J=E?new Set:new Set(Y.map(y=>y.uid)),[T,W]=V(J),z=le((y,w,e)=>{!E||e||(W(h=>{const u=new Set(h);return w?u.add(y):u.delete(y),u}),i==null||i(y,w))},[i,E]),j=y=>{const w=E&&T.has(y.uid);return $("div",{className:`pdp-downloadable-options__item ${w?"pdp-downloadable-options__item--selected":""} ${E?"":"pdp-downloadable-options__item--disabled"}`,children:[t("button",{type:"button",className:"pdp-downloadable-options__clickable","aria-pressed":w,disabled:!E,onClick:()=>E&&z(y.uid,!w,!E),onKeyDown:e=>{E&&(e.key==="Enter"||e.key===" ")&&(e.preventDefault(),z(y.uid,!w,!E))},children:t("div",{className:"pdp-downloadable-options__content",children:$("div",{className:"pdp-downloadable-options__header",children:[t("span",{className:"pdp-downloadable-options__title",children:y.label}),t(_e,{amount:y.price,currency:a,locale:n,className:"pdp-downloadable-options__price"})]})})}),y.sample_url&&t("a",{href:y.sample_url,className:"pdp-downloadable-options__sample-button",target:"_blank",rel:"noopener noreferrer","aria-label":S.label.replace("{label}",y.label),children:f})]},y.uid)},Z=y=>$("div",{className:"pdp-downloadable-options__item pdp-downloadable-options__item--sample",children:[t("button",{type:"button",className:"pdp-downloadable-options__clickable","aria-pressed":!1,disabled:!0,children:t("div",{className:"pdp-downloadable-options__content",children:t("div",{className:"pdp-downloadable-options__header",children:t("span",{className:"pdp-downloadable-options__title",children:y.label})})})}),t("a",{href:y.url,className:"pdp-downloadable-options__sample-button",target:"_blank",rel:"noopener noreferrer","aria-label":`${f} ${y.label}`,children:f})]},y.uid);return t("div",{...o,className:A(["pdp-downloadable-options",b]),"data-slot-key":"product-downloadable-options",children:$("div",{className:"pdp-downloadable-options__field",children:[t("div",{className:"pdp-downloadable-options__field__label",children:r}),$("div",{className:"pdp-downloadable-options__list",children:[Y.map(y=>j(y)),p.map(y=>Z(y))]})]})})};export{be as C,gt as D,Pt as G,ft as H,Ce as O,vt as P,bt as S,lt as Z,mt as a,wt as b,_t as c};
+//# sourceMappingURL=components.js.map
