--- scripts/__dropins__/storefront-pdp/chunks/fetchProductData.js
@@ -1,47 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{merge as De,Initializer as Oe}from"@dropins/tools/lib.js";import{events as Y}from"@dropins/tools/event-bus.js";import{PRODUCT_FRAGMENT as ye}from"../fragments.js";import{FetchGraphQL as Se}from"@dropins/tools/fetch-graphql.js";import{s as Ee}from"./getProductConfigurationValues.js";import{s as _e}from"./isProductConfigurationValid.js";function it(e){const i=new URLSearchParams(window.location.search);Object.entries(e).forEach(([n,t])=>{t===null?i.delete(n):i.set(n,String(t))});let r=window.location.pathname;r+=`?${i.toString()}`,r+=window.location.hash,window.history.replaceState({},"",r)}function ge(){const e=new URLSearchParams(window.location.search),i={};return e.forEach((r,n)=>{i[n]=r}),i}function be(e=1){const r=(ge()||{}).quantity;if(r){const n=parseInt(r,10);if(!isNaN(n)&&Number.isInteger(n)&&n>0)return n}return e}function ke(e){var i,r,n,t,a,s,l,o,m,u,f,c;return{productId:Number(e==null?void 0:e.externalId),name:e==null?void 0:e.name,sku:(e==null?void 0:e.variantSku)||(e==null?void 0:e.sku),topLevelSku:e==null?void 0:e.sku,variantName:e==null?void 0:e.variantName,specialToDate:void 0,specialFromDate:void 0,newToDate:void 0,newFromDate:void 0,createdAt:void 0,updatedAt:void 0,manufacturer:void 0,countryOfManufacture:void 0,categories:void 0,productType:e==null?void 0:e.productType,pricing:{regularPrice:((r=(i=e==null?void 0:e.prices)==null?void 0:i.regular)==null?void 0:r.amount)||0,minimalPrice:(t=(n=e==null?void 0:e.prices)==null?void 0:n.final)==null?void 0:t.minimumAmount,maximalPrice:(s=(a=e==null?void 0:e.prices)==null?void 0:a.final)==null?void 0:s.maximumAmount,specialPrice:(o=(l=e==null?void 0:e.prices)==null?void 0:l.final)==null?void 0:o.amount,tierPricing:void 0,currencyCode:((u=(m=e==null?void 0:e.prices)==null?void 0:m.final)==null?void 0:u.currency)||"USD"},canonicalUrl:e==null?void 0:e.url,mainImageUrl:(c=(f=e==null?void 0:e.images)==null?void 0:f[0])==null?void 0:c.url}}const we={PRODUCT_CONTEXT:"productContext",CHANNEL_CONTEXT:"channelContext"},ve={PRODUCT_PAGE_VIEW:"product-page-view"};function pe(){return window.adobeDataLayer=window.adobeDataLayer||[],window.adobeDataLayer}function Ae(){return{_id:"https://ns.adobe.com/xdm/channels/web",_type:"https://ns.adobe.com/xdm/channel-types/web"}}function Ce(e,i){const r=pe();r.push({[e]:null}),r.push({[e]:i})}function Ne(){typeof window>"u"||Ce(we.CHANNEL_CONTEXT,Ae())}function Ue(e,i){pe().push(n=>{const t=n.getState?n.getState():{};n.push({event:e,eventInfo:{...t,...i}})})}function Ie(e){const i=ke(e);Ce(we.PRODUCT_CONTEXT,i),Ne(),Ue(ve.PRODUCT_PAGE_VIEW)}var Z=(e=>(e.ComplexProduct="complex",e.SimpleProduct="simple",e))(Z||{}),Te=(e=>(e.ComplexProductView="ComplexProductView",e.SimpleProductView="SimpleProductView",e))(Te||{});function te(e,i){var s,l,o,m,u,f,c,T,x,O,g,I;const r=((m=(o=(l=(s=e==null?void 0:e.options)==null?void 0:s[0])==null?void 0:l.values)==null?void 0:o[0])==null?void 0:m.__typename)==="ProductViewOptionValueProduct",n=He(e,r,i==null?void 0:i.preselectFirstOption),t=(e==null?void 0:e.__typename)===Te.SimpleProductView?Z.SimpleProduct:Z.ComplexProduct,a=e?{name:e.name,sku:e.sku,isBundle:r,addToCartAllowed:e.addToCartAllowed,inStock:e.inStock,shortDescription:e.shortDescription,metaDescription:e.metaDescription,metaKeyword:e.metaKeyword,metaTitle:e.metaTitle,description:e.description,images:Fe(e),videos:Le(e),prices:qe(e,r,n),attributes:Ge(e),options:Ve(Me(e,r)),optionUIDs:n,url:e.url,urlKey:e.urlKey,externalId:e.externalId,externalParentId:e.externalParentId,variantSku:e.variantSku,variantName:e.variantName,productType:t}:null;return De(a,((T=(c=(f=(u=N.getConfig())==null?void 0:u.models)==null?void 0:f.ProductDetails)==null?void 0:c.transformer)==null?void 0:T.call(c,e))??((I=(g=(O=(x=N.getConfig())==null?void 0:x.models)==null?void 0:O.ProductDetails)==null?void 0:g.transform)==null?void 0:I.call(g,e)))}function Ve(e){var r,n,t;if(!e)return;const i=(t=(n=(r=N.getConfig())==null?void 0:r.models)==null?void 0:n.ProductOptions)==null?void 0:t.optionsTransformer;if(!i)return e;try{return i(e)??e}catch(a){return console.error("product-transform: optionsTransformer threw an error, using original options",a),e}}function Fe(e){var i,r;return(r=(i=e.images)==null?void 0:i.filter(n=>{var t;return!((t=n.roles)!=null&&t.includes("hide_from_pdp"))}))==null?void 0:r.map(n=>(n.url=n.url.replace(/^https?:/,""),n))}function Le(e){var i;return(i=e.videos)==null?void 0:i.map(r=>{var n,t;return{url:(n=r.url)==null?void 0:n.replace(/^https?:/,""),title:r.title,description:r.description,preview:r.preview?{url:(t=r.preview.url)==null?void 0:t.replace(/^https?:/,""),label:r.preview.label,roles:r.preview.roles}:void 0}})}function Ge(e){var i,r;return(r=(i=e.attributes)==null?void 0:i.filter(({roles:n})=>(n==null?void 0:n.indexOf("visible_in_pdp"))!==-1))==null?void 0:r.map(({label:n,value:t,name:a})=>{if(a!=null&&a.startsWith("ac_"))try{let s;typeof t=="string"?s=JSON.parse(t):typeof t=="object"&&(s=t);const l=R(s);return{id:a,label:n,value:JSON.stringify(l)}}catch{return{id:a,label:n,value:t.toString().split(",").join(", ")}}return{id:a,label:n,value:t.toString().split(",").join(", ")}})}function R(e){if(Array.isArray(e))return e.map(R);if(e!==null&&typeof e=="object"){const i={};for(const[r,n]of Object.entries(e)){let t=r;r==="validation_rule"?t="validationRule":r==="min_length"?t="minLength":r==="max_length"?t="maxLength":r==="purchase_separately"&&(t="purchaseSeparately"),i[t]=R(n)}return i}return e}function Me(e,i){const{options:r,optionUIDs:n}=e;return r==null?void 0:r.map(({id:t,title:a,required:s,multi:l,values:o})=>{var f,c;const m=(f=o==null?void 0:o[0])==null?void 0:f.__typename;let u=(c=o==null?void 0:o[0])==null?void 0:c.type;return u?u=u.replace("COLOR_HEX","color").replace("TEXT","text").replace("IMAGE","image"):u="dropdown",{id:t,type:u,typename:m,label:a,required:s,multiple:l,items:i?$e(o,n):Qe(o,n,u)}})}function $e(e,i){return e==null?void 0:e.filter(({enabled:r})=>typeof r>"u"||r).map(({id:r,title:n,inStock:t,isDefault:a,product:s})=>({id:r,inStock:t,label:n,selected:(i==null?void 0:i.indexOf(r))>-1||a,value:r,product:s}))}function Qe(e,i,r){return e==null?void 0:e.map(({id:n,title:t,inStock:a,value:s})=>({id:n,inStock:a,label:t,selected:(i==null?void 0:i.indexOf(n))>-1,value:(r==null?void 0:r.toLowerCase())==="dropdown"?n:s==null?void 0:s.replace(/^https?:/,"")}))}function qe(e,i,r){var I,ne,re,ie,se;const{price:n,priceRange:t,options:a,optionUIDs:s=r}=e;if(!n&&!t)return null;let{__typename:l}=e;function o(){var D,h,y,p,C,S,E,_,b,k;const w=(h=(D=n==null?void 0:n.regular)==null?void 0:D.amount)==null?void 0:h.value,P=((p=(y=n==null?void 0:n.final)==null?void 0:y.amount)==null?void 0:p.value)??((S=(C=n==null?void 0:n.regular)==null?void 0:C.amount)==null?void 0:S.value),d=((_=(E=n==null?void 0:n.regular)==null?void 0:E.amount)==null?void 0:_.currency)==="NONE"?"USD":(k=(b=n==null?void 0:n.regular)==null?void 0:b.amount)==null?void 0:k.currency;return[w,P,P,d]}function m(){var h,y,p,C,S,E,_,b,k,V,F,L,G,M,$,Q,q,H,j,X,K;const w=(p=(y=(h=t==null?void 0:t.minimum)==null?void 0:h.final)==null?void 0:y.amount)==null?void 0:p.value,P=(E=(S=(C=t==null?void 0:t.maximum)==null?void 0:C.final)==null?void 0:S.amount)==null?void 0:E.value;let d;((k=(b=(_=t==null?void 0:t.minimum)==null?void 0:_.regular)==null?void 0:b.amount)==null?void 0:k.value)===((L=(F=(V=t==null?void 0:t.maximum)==null?void 0:V.regular)==null?void 0:F.amount)==null?void 0:L.value)&&(d=($=(M=(G=t==null?void 0:t.minimum)==null?void 0:G.regular)==null?void 0:M.amount)==null?void 0:$.value);const D=((H=(q=(Q=t==null?void 0:t.minimum)==null?void 0:Q.final)==null?void 0:q.amount)==null?void 0:H.currency)==="NONE"?"USD":(K=(X=(j=t==null?void 0:t.minimum)==null?void 0:j.final)==null?void 0:X.amount)==null?void 0:K.currency;return[d,w,P,D]}function u(){var p,C,S,E,_,b,k,V,F,L,G,M,$,Q,q,H,j,X,K,ae,oe,ue,le,me;let w=0,P=0;a==null||a.forEach(U=>{var ce;const J=U==null?void 0:U.values;if(J&&Array.isArray(J)){const A=J.map(v=>{var B,W,fe,Pe;return(Pe=(fe=(W=(B=v==null?void 0:v.product)==null?void 0:B.price)==null?void 0:W.regular)==null?void 0:fe.amount)==null?void 0:Pe.value}).filter(v=>v!==void 0),z=A.length>0?Math.max(...A):0;w+=z}(ce=U==null?void 0:U.values)==null||ce.forEach(A=>{var z,v,B,W;s!=null&&s.includes(A.id)&&(P+=(W=(B=(v=(z=A==null?void 0:A.product)==null?void 0:z.price)==null?void 0:v.final)==null?void 0:B.amount)==null?void 0:W.value)})});const d=(S=(C=(p=t==null?void 0:t.minimum)==null?void 0:p.final)==null?void 0:C.amount)==null?void 0:S.value,D=(b=(_=(E=t==null?void 0:t.maximum)==null?void 0:E.final)==null?void 0:_.amount)==null?void 0:b.value;let h;((F=(V=(k=t==null?void 0:t.minimum)==null?void 0:k.regular)==null?void 0:V.amount)==null?void 0:F.value)===((M=(G=(L=t==null?void 0:t.maximum)==null?void 0:L.regular)==null?void 0:G.amount)==null?void 0:M.value)&&(h=(q=(Q=($=t==null?void 0:t.minimum)==null?void 0:$.regular)==null?void 0:Q.amount)==null?void 0:q.value);const y=((X=(j=(H=t==null?void 0:t.minimum)==null?void 0:H.final)==null?void 0:j.amount)==null?void 0:X.currency)==="NONE"?"USD":(oe=(ae=(K=t==null?void 0:t.minimum)==null?void 0:K.final)==null?void 0:ae.amount)==null?void 0:oe.currency;return i&&(s==null?void 0:s.length)<1?[h,d,D,y]:w===((me=(le=(ue=t==null?void 0:t.maximum)==null?void 0:ue.regular)==null?void 0:le.amount)==null?void 0:me.value)?[P,P,P,y]:[h,d,D,y]}const[f,c,T,x]=l==="SimpleProductView"?o():i?u():m(),O=l==="SimpleProductView"?(I=n==null?void 0:n.roles)==null?void 0:I.includes("visible"):((re=(ne=t==null?void 0:t.maximum)==null?void 0:ne.roles)==null?void 0:re.includes("visible"))&&((se=(ie=t==null?void 0:t.minimum)==null?void 0:ie.roles)==null?void 0:se.includes("visible"));function g(w){return!w||!Array.isArray(w)?[]:w.map(P=>{var d,D,h,y,p,C;return{tier:{amount:(D=(d=P.tier)==null?void 0:d.amount)==null?void 0:D.value,currency:((y=(h=P.tier)==null?void 0:h.amount)==null?void 0:y.currency)||x},quantity:((p=P.quantity[0])==null?void 0:p.gte)||((C=P.quantity[0])==null?void 0:C.lt)||0}}).sort((P,d)=>P.quantity-d.quantity)}return T&&c===T?{regular:{amount:f,currency:x,variant:f&&c!==f?"strikethrough":"default"},final:{amount:T,currency:x,variant:"default"},tiers:g(n==null?void 0:n.tiers),visible:O}:{final:{minimumAmount:c,maximumAmount:T,currency:x},tiers:g(n==null?void 0:n.tiers),visible:O}}function He(e,i,r){var n;return i?((n=e==null?void 0:e.options)==null?void 0:n.map(({values:a})=>{var o,m;const s=((o=a==null?void 0:a.find(({isDefault:u})=>u))==null?void 0:o.id)??null,l=r?(m=a[0])==null?void 0:m.id:null;return s||l})).filter(a=>a!==null):e==null?void 0:e.optionUIDs}const ee=new Oe({init:async e=>{var l,o,m;const i=(o=(l=e==null?void 0:e.models)==null?void 0:l.ProductDetails)==null?void 0:o.initialData,n=((m=(ge()||{}).optionsUIDs)==null?void 0:m.split(","))||(i==null?void 0:i.optionsUIDs),t=be(1),s={...{defaultLocale:"en-US",globalLocale:"en-US",persistURLParams:!1,acdl:!1,optionsUIDs:n},...e};if(ee.config.setConfig({...s}),s!=null&&s.sku){const u=i?te(i):await he(s.sku,{preselectFirstOption:s.preselectFirstOption});Y.emit("pdp/data",u,{scope:s.scope});const f={sku:s.sku,quantity:t,optionsUIDs:u==null?void 0:u.optionUIDs};Ee(()=>({...f}),{scope:s.scope}),_e(()=>{var c;return u!=null&&u.options?u.options.length===((c=u==null?void 0:u.optionUIDs)==null?void 0:c.length):!0},{scope:s.scope}),s.acdl&&u&&Ie(u)}},listeners:()=>[Y.on("locale",async()=>{const{sku:e,scope:i}=ee.config.getConfig();if(e){const r=await he(e);Y.emit("pdp/data",r,{scope:i})}})]}),N=ee.config,{setEndpoint:st,setFetchGraphQlHeader:at,removeFetchGraphQlHeader:ot,setFetchGraphQlHeaders:ut,getFetchGraphQlHeader:lt,fetchGraphQl:xe,getConfig:mt}=new Se().getMethods(),je=`
-query GET_PRODUCT_DATA($skus: [String]) {
-    products(skus: $skus) {
-        ...PRODUCT_FRAGMENT
-    }
-}
-
-${ye}
-`,Xe=async(e,i,r)=>{var a;const{data:n}=await xe(je,{method:"GET",variables:{skus:[e]}}),t=(a=n==null?void 0:n.products)==null?void 0:a[0];return i!=null&&i.optionsUIDs&&(t.optionUIDs=i.optionsUIDs),t?r?t:te(t,{preselectFirstOption:i==null?void 0:i.preselectFirstOption}):null},Ke=`
-query REFINE_PRODUCT_QUERY(
-    $optionIds: [String!]!
-    $sku: String!
-) {
-    # Refined Product
-    refineProduct(
-        optionIds: $optionIds 
-        sku: $sku
-    ) {
-        ...PRODUCT_FRAGMENT
-    }
-
-    # Parent Product
-    products(skus: [$sku]) {
-        ...PRODUCT_FRAGMENT
-    }
-
-    # %extendedPlaceholder%
-}
-
-${ye}
-`;async function de(e,i){const r=We(i),n=ze(r,e),t=Ke.replace("# %extendedPlaceholder%",n),{data:a}=await xe(t,{method:"GET",variables:{optionIds:i,sku:e}});return a}const Be=async(e,i,r,n)=>{var T,x,O;const t=await de(e,i);if(!t)return null;let{products:a,refineProduct:s,...l}=t;const o=a==null?void 0:a[0],m=Je(Object.values(l),o.options,r);if(r!=null&&r.length&&s===null){i=Ye(m,i,r);const g=await de(e,i);s=g==null?void 0:g.refineProduct}const u={...s||o,sku:o.sku,name:o.name,externalParentId:o==null?void 0:o.externalId,options:m,optionUIDs:i,variantSku:(s==null?void 0:s.__typename)==="SimpleProductView"?s==null?void 0:s.sku:void 0,variantName:(s==null?void 0:s.__typename)==="SimpleProductView"?s==null?void 0:s.name:void 0},f=n?u:te(u),c=(O=(x=(T=N==null?void 0:N.getConfig())==null?void 0:T.models)==null?void 0:x.ProductDetails)==null?void 0:O.fallbackData;return c?c(o,f):f};function We(e){if(e.length<2)return[e];const i=[];return e.forEach(r=>{const n=[];e.forEach(t=>{r!==t&&n.push(t)}),i.push(n)}),i}function ze(e,i){return e.map((r,n)=>`
-          ProductOption${n}: refineProduct(
-            optionIds: [
-              ${r.map(t=>`"${t}"`).join(", ")}
-            ]
-            sku: "${i}"
-          ) {
-            ... on ComplexProductView {
-              options {
-                ...PRODUCT_OPTION_FRAGMENT
-              }
-            }
-          }
-        `).join("")}function Je(e,i=[],r){const n=Object.values(e).filter(a=>!!a).reduce((a,s)=>s.options?[...a,...s.options]:[...a],[]),t=new Map(i.map(a=>[a.id,a]));return n.forEach(a=>{r!=null&&r.includes(a.id)||t.set(a.id,a)}),[...t.values()]}function Ye(e,i,r){const n=[];let t;return e.forEach(a=>{var s,l,o,m;r.includes(a.id)?t=((l=(s=a.values)==null?void 0:s.find(u=>i.includes(u==null?void 0:u.id)))==null?void 0:l.id)||((o=a.values[0])==null?void 0:o.id):t=(m=a.values[0])==null?void 0:m.id,n.push(t)}),n}const he=async(e,i)=>{const{anchors:r,optionsUIDs:n}=N.getConfig(),{optionsUIDs:t=n,anchors:a=r,preselectFirstOption:s,isBundle:l,skipTransform:o}=i??{};return!l&&t&&t.length>0?await Be(e,t,a,o):await Xe(e,{preselectFirstOption:s,optionsUIDs:t},o)};export{xe as a,st as b,N as c,at as d,ut as e,he as f,be as g,lt as h,ee as i,mt as j,Xe as k,Be as l,Ie as p,ot as r,it as s,te as t};
-//# sourceMappingURL=fetchProductData.js.map
