+++ scripts/__dropins__/storefront-pdp/containers/ProductAttributes.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsxs as u,jsx as a}from"@dropins/tools/preact-jsx-runtime.js";import{useState as f,useMemo as L,useEffect as _}from"@dropins/tools/preact-compat.js";import{classes as j,Slot as C}from"@dropins/tools/lib.js";import{events as p}from"@dropins/tools/event-bus.js";import{i as I,t as M}from"../chunks/date.js";import{g as N}from"../chunks/getFetchedProductData.js";import{useText as T}from"@dropins/tools/i18n.js";import{c as y}from"../chunks/fetchProductData.js";import"../fragments.js";import"@dropins/tools/fetch-graphql.js";import"../chunks/getProductConfigurationValues.js";import"../chunks/isProductConfigurationValid.js";const F=({initialData:b=null,scope:i,children:H,slots:e,formatValue:c,...m})=>{var s;const d=T("PDP.Product.Details.label").label,{globalLocale:g="en-US"}=y.getConfig(),[D,h]=f(g),[r,S]=f(b),x=L(()=>({data:r}),[r]);return _(()=>{const t=p.on("pdp/data",S,{scope:i}),o=p.on("locale",h);return()=>{t==null||t.off(),o==null||o.off()}},[i]),r?u("div",{...m,className:j(["pdp-attributes",m.className]),children:[d,e!=null&&e.Attributes?a(C,{name:"Attributes",slot:e==null?void 0:e.Attributes,context:{...x}}):a("ul",{children:(s=r==null?void 0:r.attributes)==null?void 0:s.map(({id:t,label:o,value:n},P)=>{if(t!=null&&t.startsWith("ac_")||!n)return null;const l=I(n)?M(n,D):n.toString(),A=c?c(l,t,o):l;return u("li",{children:[o,": ",a("span",{dangerouslySetInnerHTML:{__html:A}})]},`attribute-${P}`)})})]}):null};F.getInitialData=N;export{F as ProductAttributes,F as default};
+import{jsxs as m,jsx as a}from"@dropins/tools/preact-jsx-runtime.js";import{useState as f,useMemo as L,useEffect as _}from"@dropins/tools/preact-compat.js";import{Slot as j,classes as C}from"@dropins/tools/lib.js";import{events as p}from"@dropins/tools/event-bus.js";import{i as I,t as M}from"../chunks/date.js";import{config as N,getFetchedProductData as T}from"../api.js";import{useText as y}from"@dropins/tools/i18n.js";import"@dropins/tools/fetch-graphql.js";const F=({initialData:b=null,scope:i,children:H,slots:e,formatValue:c,...s})=>{var l;const d=y("PDP.Product.Details.label").label,{globalLocale:g="en-US"}=N.getConfig(),[D,h]=f(g),[r,S]=f(b),x=L(()=>({data:r}),[r]);return _(()=>{const t=p.on("pdp/data",S,{scope:i}),o=p.on("locale",h);return()=>{t==null||t.off(),o==null||o.off()}},[i]),r?m("div",{...s,className:C(["pdp-attributes",s.className]),children:[d,e!=null&&e.Attributes?a(j,{name:"Attributes",slot:e==null?void 0:e.Attributes,context:{...x}}):a("ul",{children:(l=r==null?void 0:r.attributes)==null?void 0:l.map(({id:t,label:o,value:n},P)=>{if(t!=null&&t.startsWith("ac_")||!n)return null;const u=I(n)?M(n,D):n.toString(),A=c?c(u,t,o):u;return m("li",{children:[o,": ",a("span",{dangerouslySetInnerHTML:{__html:A}})]},`attribute-${P}`)})})]}):null};F.getInitialData=T;export{F as ProductAttributes,F as default};
 //# sourceMappingURL=ProductAttributes.js.map
