+++ scripts/__dropins__/storefront-pdp/containers/ProductDescription.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as i}from"@dropins/tools/preact-jsx-runtime.js";import{useState as a,useEffect as c}from"@dropins/tools/preact-compat.js";import{classes as m}from"@dropins/tools/lib.js";import{events as l}from"@dropins/tools/event-bus.js";import{g as f}from"../chunks/getFetchedProductData.js";const u=({initialData:s=null,scope:r,children:d,...o})=>{const[t,n]=a(s);return c(()=>{const e=l.on("pdp/data",n,{scope:r});return()=>{e==null||e.off()}},[r]),t?i("div",{dangerouslySetInnerHTML:{__html:(t==null?void 0:t.description)??""},...o,className:m(["pdp-description",o.className])}):null};u.getInitialData=f;export{u as ProductDescription,u as default};
+import{jsx as n}from"@dropins/tools/preact-jsx-runtime.js";import{useState as a,useEffect as m}from"@dropins/tools/preact-compat.js";import{classes as c}from"@dropins/tools/lib.js";import{events as l}from"@dropins/tools/event-bus.js";import{getFetchedProductData as f}from"../api.js";import"@dropins/tools/fetch-graphql.js";const p=({initialData:s=null,scope:e,children:u,...o})=>{const[t,i]=a(s);return m(()=>{const r=l.on("pdp/data",i,{scope:e});return()=>{r==null||r.off()}},[e]),t?n("div",{dangerouslySetInnerHTML:{__html:(t==null?void 0:t.description)??""},...o,className:c(["pdp-description",o.className])}):null};p.getInitialData=f;export{p as ProductDescription,p as default};
 //# sourceMappingURL=ProductDescription.js.map
