+++ scripts/__dropins__/storefront-pdp/containers/ProductDetails.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as e,jsxs as B,Fragment as Jt}from"@dropins/tools/preact-jsx-runtime.js";import{useState as J,Children as qt,useEffect as dt,useCallback as ct,useMemo as Zt}from"@dropins/tools/preact-compat.js";import{VComponent as m,classes as N,Slot as P,toLanguageTag as Gt}from"@dropins/tools/lib.js";import{Incrementer as $t,Image as Rt,Button as Tt,Icon as yt,Breadcrumbs as Ot}from"@dropins/tools/components.js";/* empty css                  */import{Z as Kt,C as lt,O as Wt}from"../chunks/Zoom.js";import"@dropins/tools/preact-hooks.js";/* empty css                       *//* empty css                           */import{useText as G}from"@dropins/tools/i18n.js";import{P as jt}from"../chunks/PriceRange.js";import{S as Xt}from"../chunks/Swatches.js";import"@dropins/tools/preact.js";import{G as Yt}from"../chunks/GiftCardOptions2.js";import{D as ft}from"../chunks/DownloadableOptions2.js";import{c as Ht,g as gt,f as Nt,p as Ct,s as zt,t as te}from"../chunks/fetchProductData.js";import{events as it}from"@dropins/tools/event-bus.js";import{i as ee,t as ne}from"../chunks/date.js";import"../fragments.js";import"@dropins/tools/fetch-graphql.js";import"../chunks/getProductConfigurationValues.js";import"../chunks/isProductConfigurationValid.js";const re=({title:u,breadcrumbs:I,galleryContent:S,infoContent:w,productContent:V,sku:U,outOfStock:E,shortDescription:R,description:p,attributes:ot,images:c,options:o,giftCardOptions:s,downloadableOptions:C,price:Q,specialPrice:tt,className:at,children:Z,quantity:et,actions:$,carouselConfig:n,zoomType:ht="overlay",closeButton:ut=!0,...L})=>{var T,_,W,q,nt,rt,X,j,pt;const mt=G("PDP.Product.RegularPrice.label").label,_t=G("PDP.Product.SpecialPrice.label").label,st=G("PDP.Product.OutOfStock.label").label,t=G("PDP.Product.Image.label").label,[M,A]=J(null),k=c!=null&&c.length?c.map((F,O)=>{var Y;const z=((Y=u==null?void 0:u.props)==null?void 0:Y.children)&&(t==null?void 0:t.replace("{product}",u.props.children.toString()).replace("{key}",String(O+1)).replace("{total}",String(c.length)));return ht==="zoom"?e(Kt,{closeButton:ut,children:e(m,{node:F,loading:O===0?"eager":"lazy",alt:z},O)}):e(m,{node:F,loading:O===0?"eager":"lazy",alt:z,onClick:()=>A(O)},O)}):e("img",{src:"data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7",className:"pdp-product__images__placeholder","aria-hidden":!0}),K=(T=n==null?void 0:n.thumbnails)!=null&&T.length?n.thumbnails.map((F,O)=>{var z;return e(m,{node:F,loading:(n==null?void 0:n.thumbnailsLoadingMode)||"lazy",alt:((z=u==null?void 0:u.props)==null?void 0:z.children)&&(t==null?void 0:t.replace("{product}",u.props.children.toString()).replace("{key}",String(O+1)).replace("{total}",String(c.length)))},O)}):e("img",{src:"data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7",className:"pdp-product__images__placeholder","aria-hidden":!0});return B("div",{...L,className:N(["pdp-product",at]),children:[I&&e(m,{node:I,className:"pdp-product__breadcrumbs"}),B("div",{className:"pdp-product__column-container pdp-product__gallery-column",children:[e("div",{className:"pdp-product__column-body",children:(_=n==null?void 0:n.controls)!=null&&_.desktop?e(lt,{className:"pdp-product__overlay__carousel",arrows:(c==null?void 0:c.length)>1,controls:(c==null?void 0:c.length)>1?n.controls.desktop:"dots",arrowsOnMainImage:n==null?void 0:n.arrowsOnMainImage,loop:(n==null?void 0:n.loopable)&&(c==null?void 0:c.length)>1,peak:((W=n==null?void 0:n.thumbnails)==null?void 0:W.length)>1&&((q=n==null?void 0:n.peak)==null?void 0:q.desktop)||!1,gap:(n==null?void 0:n.gap)||null,width:"100%",height:"100%",defaultIndex:M||0,thumbnails:K,children:k}):e(pe,{className:"pdp-product__images",gap:"small",children:k})}),S&&e(m,{node:S,className:"pdp-product__gallery-content"})]}),B("div",{className:"pdp-product__column-container pdp-product__content-column",children:[B("div",{className:"pdp-product__column-body",children:[B("div",{className:"pdp-product__header",children:[u&&e(m,{node:u,className:"pdp-product__title"}),U&&e(m,{node:U,className:"pdp-product__sku"})]}),B("div",{className:"pdp-product__prices",children:[tt&&e(m,{node:tt,className:N(["pdp-product__price-special","pdp-product__price"]),"aria-label":_t,role:"text"}),Q&&e(m,{node:Q,className:N(["pdp-product__price-regular","pdp-product__price--grey"]),"aria-label":mt,role:"text"})]}),(nt=n==null?void 0:n.controls)!=null&&nt.mobile&&((rt=n==null?void 0:n.controls)==null?void 0:rt.mobile)==="thumbnailsRow"?e(lt,{className:"pdp-product__images pdp-product__images--carousel pdp-product__images--carousel--thumbnails",width:"100%",height:"auto",arrows:(c==null?void 0:c.length)>1,controls:(c==null?void 0:c.length)>1?"thumbnailsRow":null,loop:(n==null?void 0:n.loopable)&&(c==null?void 0:c.length)>1,peak:((X=n==null?void 0:n.thumbnails)==null?void 0:X.length)>1&&((j=n==null?void 0:n.peak)==null?void 0:j.mobile)||!1,gap:(n==null?void 0:n.gap)||null,thumbnails:K,children:k}):e(lt,{className:"pdp-product__images pdp-product__images--carousel",peak:((pt=n==null?void 0:n.peak)==null?void 0:pt.mobile)&&(c==null?void 0:c.length)>1,gap:(n==null?void 0:n.gap)||null,width:"100%",height:"auto",arrows:(c==null?void 0:c.length)>1,controls:(c==null?void 0:c.length)>1?"dots":null,loop:(n==null?void 0:n.loopable)&&(c==null?void 0:c.length)>1,children:k}),B("div",{className:N(["pdp-product__actions",["pdp-product__actions--out-of-stock",E]]),children:[E&&e("div",{className:"pdp-product__out-of-stock__text",children:st}),o&&e(m,{node:o,className:N(["pdp-product__options"])}),s&&e(m,{node:s,className:N(["pdp-product__gift-card-options"])}),C&&e(m,{node:C,className:N(["pdp-product__downloadable-options"])}),et&&e("div",{className:"pdp-product__quantity",children:e(m,{node:et})}),$&&e(m,{node:$,className:N(["pdp-product__buttons"])})]}),R&&e(m,{node:R,className:N(["pdp-product__short_description"])}),p&&e(m,{node:p,className:N(["pdp-product__description"])}),ot?e("div",{className:"pdp-product__attributes",children:ot}):null]}),w&&e(m,{node:w,className:"pdp-product__info-content"})]}),V&&e(m,{node:V,className:"pdp-product__content"}),M===null?null:e(Wt,{className:"pdp-product__overlay",onClose:()=>A(null),centered:!0,children:e(lt,{className:"pdp-product__overlay__carousel",arrows:(c==null?void 0:c.length)>1,width:"100%",height:"100%",defaultIndex:M||0,loop:(n==null?void 0:n.loopable)&&(c==null?void 0:c.length)>1,controls:(c==null?void 0:c.length)>1?"dots":null,isZoomed:!0,children:k})})]})},pe=({children:u,gap:I=null,className:S,style:w,...V})=>{const U=qt.count(u);return e("div",{...V,className:N(["pdp-gallery-grid",S]),style:{"--gap":I?`var(--spacing-${I})`:"0px",...w},children:qt.map(u,(E,R)=>e("div",{className:"pdp-gallery-grid__item",children:e(m,{node:E,loading:R===0?"eager":"lazy"},E.props.src)},U+R))})},de=({sku:u,hideSku:I,hideQuantity:S,hideShortDescription:w,hideDescription:V,hideAttributes:U,hideURLParams:E,hideSelectedOptionValue:R,slots:p,children:ot,initialData:c,carousel:o,optionsConfig:s,useACDL:C,onAddToCart:Q,zoomType:tt,closeButton:at=!0,disableDropdownPreselection:Z,...et})=>{var Y,St,Dt,It,wt,Vt,Ut,kt,xt,Bt,Et;const $=G("PDP.Product.AddToCart.label").label,n=G("PDP.Product.Incrementer.label").label,ht=G("PDP.Product.Details.label").label,ut=G("PDP.Swatches.ChooseOption.label").label,[L,mt]=J(Ht.getConfig().defaultLocale||"en-US"),[_t,st]=J(),[t=null,M]=J(c),[A,k]=J(()=>{const r=gt(1),d={sku:u,quantity:r};return t!=null&&t.optionUIDs&&(d.optionsUIDs=t.optionUIDs),d}),[K,T]=J(()=>{var r,d;return!(t!=null&&t.options&&((r=t==null?void 0:t.options)==null?void 0:r.length)>0&&((d=t==null?void 0:t.optionUIDs)==null?void 0:d.length)!==(t==null?void 0:t.options.length))}),_={data:t,values:A,valid:K};function W(r){return r==null?void 0:r.some(d=>(d==null?void 0:d.typename)==="ProductViewOptionValueProduct")}dt(()=>{const r=it.on("pdp/data",l=>{l&&(M(l),k(i=>({...i,sku:l.sku,optionsUIDs:l.optionUIDs})))}),d=it.on("pdp/setValues",l=>{k(i=>({...i,...l}))});return()=>{d==null||d.off(),r==null||r.off()}},[]),dt(()=>{const r=it.on("locale",d=>{d!==L&&(mt(d),Nt(u,{preselectFirstOption:!Z,anchors:s==null?void 0:s.anchorOptions}).then(M))});return()=>{r==null||r.off()}},[Z,L,s==null?void 0:s.anchorOptions,u]),dt(()=>{C&&c&&Object.keys(c).length>0&&Ct(c)},[c]);const q=ct(r=>{k(d=>({...d,...r}))},[]);dt(()=>{it.emit("pdp/values",A)},[JSON.stringify(A)]);const nt=ct(async r=>{var i,a;let d=Object.entries(r).filter(([h,b])=>b.label!==ut).map(([h,b])=>r[h].value).filter(h=>h!==void 0);!E&&(t!=null&&t.options)&&!W(t.options)&&zt({optionsUIDs:d.join(",")}),q({optionsUIDs:d}),T((d==null?void 0:d.length)===((i=t==null?void 0:t.options)==null?void 0:i.length));const l=await Nt(u,{anchors:s==null?void 0:s.anchorOptions,isBundle:t==null?void 0:t.isBundle,optionsUIDs:d,preselectFirstOption:!Z});M(l),(a=s==null?void 0:s.anchorOptions)!=null&&a.length&&rt(l,d)},[u]),rt=ct((r,d)=>{var a;const l=(r==null?void 0:r.optionUIDs)??[];d.every(h=>l.includes(h))||(st(r==null?void 0:r.options),q({optionsUIDs:l}),T((l==null?void 0:l.length)===((a=r==null?void 0:r.options)==null?void 0:a.length)))},[u]),X=ct(r=>{Object.keys(r).length!==0&&T(!1)},[]),j=Zt(()=>{var r;return(r=t==null?void 0:t.attributes)==null?void 0:r.map(({label:d,value:l,id:i},a)=>{if(i==="ac_giftcard")return"";const h=ee(l)?ne(l,L):l.toString();return B("li",{children:[d,": ",e("span",{dangerouslySetInnerHTML:{__html:h}})]},a)})},[t==null?void 0:t.attributes,L]),pt=!S&&(t!=null&&t.inStock)?e(P,{name:"Quantity",slot:p==null?void 0:p.Quantity,context:{..._},children:e($t,{name:"quantity",defaultValue:A.quantity.toString(),value:A.quantity.toString(),min:1,"aria-label":n,className:"pdp-product__quantity",onValue:r=>{const d=Number(r);q({quantity:d}),!E&&(t!=null&&t.options)&&(W(t.options)||zt({quantity:d}))}},"quantity")}):void 0,F=()=>{if(t!=null&&t.options)return e(P,{name:"Options",slot:p==null?void 0:p.Options,context:{..._},children:e(Xt,{selectedUIDs:A.optionsUIDs,options:t.options,defaultOptions:A.optionsUIDs,selectionsToUpdate:_t,hideSelectedValue:R,disablePreselections:Z,onValues:nt,onErrors:X})})},O=()=>{if(t!=null&&t.attributes){const r=t.attributes.filter(d=>d.id==="ac_giftcard");if(r.length>0){const l=r.map((i,a)=>{var h;try{const b=JSON.parse(i.value),H=A.enteredOptions||[],f=((h=b.options)==null?void 0:h.map(v=>{const y=H.find(D=>D.uid===v.uid);return{id:v.uid,value:(y==null?void 0:y.value)||""}}))||[];return e(Yt,{options:b.options||[],optionValues:f,onOptionChange:(v,y)=>{var Lt;const D=A.enteredOptions||[],bt=D.findIndex(x=>x.uid===v);let At;if(bt!==-1?At=D.map(x=>x.uid===v?{...x,value:y}:x):At=[...D,{uid:v,value:y}],q({enteredOptions:At}),t){const x={...t},Qt=((Lt=x.attributes)==null?void 0:Lt.map(g=>{var Mt;if(g.id==="ac_giftcard")try{const Pt=JSON.parse(g.value),Ft=((Mt=Pt.options)==null?void 0:Mt.map(vt=>vt.uid===v?{...vt,value:y}:vt))||[];return{...g,value:JSON.stringify({...Pt,options:Ft})}}catch{return g}return g}))||[];x.attributes=Qt,M(x)}},onValidationChange:v=>{v||X({giftCard:"Invalid gift card data"}),T(v)},className:"pdp-product__gift-card-options"},`giftcard-${a}`)}catch(b){return console.warn("Invalid gift card data format:",b),null}}).filter(i=>i!==null);return l.length>0?e("div",{className:"pdp-product__gift-card-options-container",children:l}):void 0}}},z=()=>{if(t!=null&&t.attributes){const r=t.attributes.filter(d=>{var l;return d.id==="ac_downloadable"&&((l=d.roles)==null?void 0:l.includes("visible_in_pdp"))});if(r.length>0){const l=r.map((i,a)=>{var h,b;try{const H=typeof i.value=="string"?JSON.parse(i.value):i.value;return e(ft,{label:i.label,value:H,onLinkChange:(f,v)=>{const y=A.optionsUIDs||[];let D;v?y.includes(f)?D=y:D=[...y,f]:D=y.filter(bt=>bt!==f),q({optionsUIDs:D})},currency:(b=(h=t.prices)==null?void 0:h.final)==null?void 0:b.currency,className:"pdp-product__downloadable-options"},`downloadable-${a}`)}catch(H){return console.warn("Invalid downloadable data format:",H),null}}).filter(i=>i!==null);return l.length>0?e("div",{className:"pdp-product__downloadable-options-container",children:l}):void 0}}};return e("div",{...et,children:e(re,{title:e(P,{name:"Title",slot:p==null?void 0:p.Title,context:{..._},children:t==null?void 0:t.name}),sku:I?void 0:e(P,{name:"SKU",slot:p==null?void 0:p.SKU,context:{..._},children:t==null?void 0:t.sku}),options:F(),giftCardOptions:O(),downloadableOptions:z(),breadcrumbs:(p==null?void 0:p.Breadcrumbs)&&e(P,{name:"Breadcrumbs",slot:p==null?void 0:p.Breadcrumbs,context:{..._,setSeparator(r){this._registerMethod((...d)=>{const l=r(...d),i=l&&e(yt,{source:l});this._setProps(a=>({...a,separator:i,children:[e(Ot,{separator:i,categories:a.categories})]}))})},appendLink(r){this._registerMethod((...d)=>{const{text:l,...i}=r(...d),a=e("a",{...i,children:l});this._setProps(h=>{const b=[...h.categories||[],a];return{...h,categories:b,children:[e(Ot,{separator:h.separator,categories:b})]}})})},appendHTMLElement(r){this._registerMethod((...d)=>{const l=r(...d),i=this._htmlElementToVNode(l);this._setProps(a=>{const h=[...a.categories||[],i];return{...a,categories:h,children:[e(Ot,{separator:a.separator,categories:h})]}})})}}}),quantity:pt,actions:e(P,{name:"Actions",slot:p==null?void 0:p.Actions,context:{..._,appendButton(r){this._registerMethod((...d)=>{const l=r(...d);if(!l)return;const{text:i,icon:a,...h}=l,b=e(Tt,{type:"button",...h,icon:a&&e(yt,{source:a}),children:i});this._setProps(H=>({children:[...H.children||[],b]}))})}},children:!(p!=null&&p.Actions)&&e(Tt,{size:"medium",type:"submit",icon:e(yt,{source:"Cart"}),disabled:!(t!=null&&t.inStock)||!K,"aria-label":$,onClick:()=>Q==null?void 0:Q(A),children:$})}),shortDescription:w?void 0:e(P,{name:"ShortDescription",slot:p==null?void 0:p.ShortDescription,context:{..._},children:e("div",{dangerouslySetInnerHTML:{__html:(t==null?void 0:t.shortDescription)??""}})}),description:V?void 0:e(P,{name:"Description",slot:p==null?void 0:p.Description,context:{..._},children:e("div",{dangerouslySetInnerHTML:{__html:(t==null?void 0:t.description)??""}})}),images:((Y=t==null?void 0:t.images)==null?void 0:Y.map(({label:r,url:d})=>{var l,i;return e(Rt,{title:r,alt:r,src:d,width:((l=o==null?void 0:o.imageParams)==null?void 0:l.width)??960,height:((i=o==null?void 0:o.imageParams)==null?void 0:i.height)??1191,params:(o==null?void 0:o.imageParams)??{width:960}},d)}))??[],specialPrice:(St=t==null?void 0:t.prices)!=null&&St.visible?e(P,{name:"SpecialPrice",slot:p==null?void 0:p.SpecialPrice,context:{..._},children:e(jt,{...(Dt=t.prices)==null?void 0:Dt.final,locale:Gt(L)})}):void 0,price:(It=t==null?void 0:t.prices)!=null&&It.visible&&((wt=t==null?void 0:t.prices)!=null&&wt.regular)&&((Ut=(Vt=t==null?void 0:t.prices)==null?void 0:Vt.final)==null?void 0:Ut.amount)!==((xt=(kt=t==null?void 0:t.prices)==null?void 0:kt.regular)==null?void 0:xt.amount)?e(P,{name:"RegularPrice",slot:p==null?void 0:p.RegularPrice,context:{..._},children:e(jt,{...(Bt=t.prices)==null?void 0:Bt.regular,locale:Gt(L)})}):void 0,zoomType:tt,closeButton:at,carouselConfig:{...o,thumbnails:o&&Object.keys(o).length?((Et=t==null?void 0:t.images)==null?void 0:Et.map(({label:r,url:d})=>{var l,i;return e(Rt,{title:r,alt:r,src:d,width:((l=o==null?void 0:o.thumbnailParams)==null?void 0:l.width)??200,height:((i=o==null?void 0:o.thumbnailParams)==null?void 0:i.height)??248,params:(o==null?void 0:o.thumbnailParams)??{width:200}},d)}))??[]:[]},outOfStock:!(t!=null&&t.inStock),attributes:U?void 0:e(P,{name:"Attributes",slot:p==null?void 0:p.Attributes,context:{..._},children:j!=null&&j.length?B(Jt,{children:[ht,e("ul",{children:j})]}):null}),galleryContent:(p==null?void 0:p.GalleryContent)&&e(P,{name:"GalleryContent",slot:p.GalleryContent,context:{..._}}),infoContent:(p==null?void 0:p.InfoContent)&&e(P,{name:"InfoContent",slot:p.InfoContent,context:{..._}}),productContent:(p==null?void 0:p.Content)&&e(P,{name:"Content",slot:p.Content,context:{..._}})})})};de.getInitialData=async function({sku:u,disableDropdownPreselection:I=!1,optionsConfig:S}){var V,U;const w=(U=(V=Ht.getConfig().models)==null?void 0:V.ProductDetails)==null?void 0:U.initialData;return w?te(w):Nt(u,{preselectFirstOption:!I,anchors:S==null?void 0:S.anchorOptions})};export{de as ProductDetails,de as default};
+import{jsxs as he,jsx as o,Fragment as Fe}from"@dropins/tools/preact-jsx-runtime.js";import{useState as V,useEffect as C,useCallback as M,useMemo as Je}from"@dropins/tools/preact-compat.js";import{Slot as u,toLanguageTag as fe}from"@dropins/tools/lib.js";import{Incrementer as He,Image as be,Button as ge,Icon as K,Breadcrumbs as R}from"@dropins/tools/components.js";import{c as Qe,a as Oe,D as Ae,G as ze,S as Ke}from"../chunks/components.js";import"@dropins/tools/preact.js";import{config as Pe,g as Re,i as $,fetchProductData as W,p as $e,s as We,c as De,t as Xe}from"../api.js";import{events as T}from"@dropins/tools/event-bus.js";import{i as Ye,t as Ze}from"../chunks/date.js";import{useText as j}from"@dropins/tools/i18n.js";import"@dropins/tools/preact-hooks.js";import"@dropins/tools/fetch-graphql.js";const xe=({sku:D,hideSku:G,hideQuantity:I,hideShortDescription:w,hideDescription:E,hideAttributes:L,hideURLParams:X,hideSelectedOptionValue:Se,slots:n,children:et,initialData:y,carousel:s,optionsConfig:d,useACDL:Ie,onAddToCart:F,zoomType:ye,closeButton:ve=!0,disableDropdownPreselection:N,..._e})=>{var te,ne,ie,re,ce,oe,ae,se,pe,le,ue;const Y=j("PDP.Product.AddToCart.label").label,Ue=j("PDP.Product.Incrementer.label").label,Ve=j("PDP.Product.Details.label").label,we=j("PDP.Swatches.ChooseOption.label").label,[S,Ee]=V(Pe.getConfig().defaultLocale||"en-US"),[Le,Ne]=V(),[e=null,k]=V(y),[b,J]=V(()=>{const t=Re(1),i={sku:D,quantity:t};return e!=null&&e.optionUIDs&&(i.optionsUIDs=e.optionUIDs),i}),[Z,B]=V(()=>{var t;return(t=e==null?void 0:e.options)!=null&&t.length?$(e.options,e.optionUIDs,e.isBundle):!0}),l={data:e,values:b,valid:Z};function x(t){return t==null?void 0:t.some(i=>(i==null?void 0:i.typename)==="ProductViewOptionValueProduct")}C(()=>{const t=T.on("pdp/data",r=>{r&&(k(r),J(c=>({...c,sku:r.sku,optionsUIDs:r.optionUIDs})))}),i=T.on("pdp/setValues",r=>{J(c=>({...c,...r}))});return()=>{i==null||i.off(),t==null||t.off()}},[]),C(()=>{const t=T.on("locale",i=>{i!==S&&(Ee(i),W(D,{preselectFirstOption:!N,anchors:d==null?void 0:d.anchorOptions}).then(k))});return()=>{t==null||t.off()}},[N,S,d==null?void 0:d.anchorOptions,D]),C(()=>{Ie&&y&&Object.keys(y).length>0&&$e(y)},[y]);const v=M(t=>{J(i=>({...i,...t}))},[]);C(()=>{T.emit("pdp/values",b)},[JSON.stringify(b)]);const ke=M(async t=>{var a;const i=Object.fromEntries(Object.entries(t).filter(([,p])=>p.label!==we));let r=We(i,e==null?void 0:e.options);!X&&(e!=null&&e.options)&&!x(e.options)&&De({optionsUIDs:r.join(",")}),v({optionsUIDs:r});const c=await W(D,{anchors:d==null?void 0:d.anchorOptions,isBundle:e==null?void 0:e.isBundle,optionsUIDs:r,preselectFirstOption:!N});k(c),B($(c==null?void 0:c.options,(c==null?void 0:c.optionUIDs)??r,c==null?void 0:c.isBundle)),(a=d==null?void 0:d.anchorOptions)!=null&&a.length&&Be(c,r)},[D]),Be=M((t,i)=>{const r=(t==null?void 0:t.optionUIDs)??[];i.every(a=>r.includes(a))||(Ne(t==null?void 0:t.options),v({optionsUIDs:r}),B($(t==null?void 0:t.options,r,t==null?void 0:t.isBundle)))},[D]),ee=M(t=>{Object.keys(t).length!==0&&B(!1)},[]),q=Je(()=>{var t;return(t=e==null?void 0:e.attributes)==null?void 0:t.map(({label:i,value:r,id:c},a)=>{if(c==="ac_giftcard")return"";const p=Ye(r)?Ze(r,S):r.toString();return he("li",{children:[i,": ",o("span",{dangerouslySetInnerHTML:{__html:p}})]},a)})},[e==null?void 0:e.attributes,S]),qe=!I&&(e!=null&&e.inStock)?o(u,{name:"Quantity",slot:n==null?void 0:n.Quantity,context:{...l},children:o(He,{name:"quantity",defaultValue:b.quantity.toString(),value:b.quantity.toString(),min:1,"aria-label":Ue,className:"pdp-product__quantity",onValue:t=>{const i=Number(t);v({quantity:i}),!X&&(e!=null&&e.options)&&(x(e.options)||De({quantity:i}))}},"quantity")}):void 0,Ce=()=>{if(e!=null&&e.options)return o(u,{name:"Options",slot:n==null?void 0:n.Options,context:{...l},children:o(Ke,{selectedUIDs:b.optionsUIDs,options:e.options,defaultOptions:b.optionsUIDs,selectionsToUpdate:Le,hideSelectedValue:Se,disablePreselections:N,onValues:ke,onErrors:ee})})},Me=()=>{if(e!=null&&e.attributes){const t=e.attributes.filter(i=>i.id==="ac_giftcard");if(t.length>0){const r=t.map((c,a)=>{var p;try{const m=JSON.parse(c.value),P=b.enteredOptions||[],_=((p=m.options)==null?void 0:p.map(h=>{const f=P.find(g=>g.uid===h.uid);return{id:h.uid,value:(f==null?void 0:f.value)||""}}))||[];return o(ze,{options:m.options||[],optionValues:_,onOptionChange:(h,f)=>{var de;const g=b.enteredOptions||[],H=g.findIndex(O=>O.uid===h);let Q;if(H!==-1?Q=g.map(O=>O.uid===h?{...O,value:f}:O):Q=[...g,{uid:h,value:f}],v({enteredOptions:Q}),e){const O={...e},je=((de=O.attributes)==null?void 0:de.map(U=>{var me;if(U.id==="ac_giftcard")try{const A=JSON.parse(U.value),Ge=((me=A.options)==null?void 0:me.map(z=>z.uid===h?{...z,value:f}:z))||[];return{...U,value:JSON.stringify({...A,options:Ge})}}catch{return U}return U}))||[];O.attributes=je,k(O)}},onValidationChange:h=>{h||ee({giftCard:"Invalid gift card data"}),B(h)},className:"pdp-product__gift-card-options"},`giftcard-${a}`)}catch(m){return console.warn("Invalid gift card data format:",m),null}}).filter(c=>c!==null);return r.length>0?o("div",{className:"pdp-product__gift-card-options-container",children:r}):void 0}}},Te=()=>{if(e!=null&&e.attributes){const t=e.attributes.filter(i=>{var r;return i.id==="ac_downloadable"&&((r=i.roles)==null?void 0:r.includes("visible_in_pdp"))});if(t.length>0){const r=t.map((c,a)=>{var p,m;try{const P=typeof c.value=="string"?JSON.parse(c.value):c.value;return o(Ae,{label:c.label,value:P,onLinkChange:(_,h)=>{const f=b.optionsUIDs||[];let g;h?f.includes(_)?g=f:g=[...f,_]:g=f.filter(H=>H!==_),v({optionsUIDs:g})},currency:(m=(p=e.prices)==null?void 0:p.final)==null?void 0:m.currency,className:"pdp-product__downloadable-options"},`downloadable-${a}`)}catch(P){return console.warn("Invalid downloadable data format:",P),null}}).filter(c=>c!==null);return r.length>0?o("div",{className:"pdp-product__downloadable-options-container",children:r}):void 0}}};return o("div",{..._e,children:o(Qe,{title:o(u,{name:"Title",slot:n==null?void 0:n.Title,context:{...l},children:e==null?void 0:e.name}),sku:G?void 0:o(u,{name:"SKU",slot:n==null?void 0:n.SKU,context:{...l},children:e==null?void 0:e.sku}),options:Ce(),giftCardOptions:Me(),downloadableOptions:Te(),breadcrumbs:(n==null?void 0:n.Breadcrumbs)&&o(u,{name:"Breadcrumbs",slot:n==null?void 0:n.Breadcrumbs,context:{...l,setSeparator(t){this._registerMethod((...i)=>{const r=t(...i),c=r&&o(K,{source:r});this._setProps(a=>({...a,separator:c,children:[o(R,{separator:c,categories:a.categories})]}))})},appendLink(t){this._registerMethod((...i)=>{const{text:r,...c}=t(...i),a=o("a",{...c,children:r});this._setProps(p=>{const m=[...p.categories||[],a];return{...p,categories:m,children:[o(R,{separator:p.separator,categories:m})]}})})},appendHTMLElement(t){this._registerMethod((...i)=>{const r=t(...i),c=this._htmlElementToVNode(r);this._setProps(a=>{const p=[...a.categories||[],c];return{...a,categories:p,children:[o(R,{separator:a.separator,categories:p})]}})})}}}),quantity:qe,actions:o(u,{name:"Actions",slot:n==null?void 0:n.Actions,context:{...l,appendButton(t){this._registerMethod((...i)=>{const r=t(...i);if(!r)return;const{text:c,icon:a,...p}=r,m=o(ge,{type:"button",...p,icon:a&&o(K,{source:a}),children:c});this._setProps(P=>({children:[...P.children||[],m]}))})}},children:!(n!=null&&n.Actions)&&o(ge,{size:"medium",type:"submit",icon:o(K,{source:"Cart"}),disabled:!(e!=null&&e.inStock)||!Z,"aria-label":Y,onClick:()=>F==null?void 0:F(b),children:Y})}),shortDescription:w?void 0:o(u,{name:"ShortDescription",slot:n==null?void 0:n.ShortDescription,context:{...l},children:o("div",{dangerouslySetInnerHTML:{__html:(e==null?void 0:e.shortDescription)??""}})}),description:E?void 0:o(u,{name:"Description",slot:n==null?void 0:n.Description,context:{...l},children:o("div",{dangerouslySetInnerHTML:{__html:(e==null?void 0:e.description)??""}})}),images:((te=e==null?void 0:e.images)==null?void 0:te.map(({label:t,url:i})=>{var r,c;return o(be,{title:t,alt:t,src:i,width:((r=s==null?void 0:s.imageParams)==null?void 0:r.width)??960,height:((c=s==null?void 0:s.imageParams)==null?void 0:c.height)??1191,params:(s==null?void 0:s.imageParams)??{width:960}},i)}))??[],specialPrice:(ne=e==null?void 0:e.prices)!=null&&ne.visible?o(u,{name:"SpecialPrice",slot:n==null?void 0:n.SpecialPrice,context:{...l},children:o(Oe,{...(ie=e.prices)==null?void 0:ie.final,locale:fe(S)})}):void 0,price:(re=e==null?void 0:e.prices)!=null&&re.visible&&((ce=e==null?void 0:e.prices)!=null&&ce.regular)&&((ae=(oe=e==null?void 0:e.prices)==null?void 0:oe.final)==null?void 0:ae.amount)!==((pe=(se=e==null?void 0:e.prices)==null?void 0:se.regular)==null?void 0:pe.amount)?o(u,{name:"RegularPrice",slot:n==null?void 0:n.RegularPrice,context:{...l},children:o(Oe,{...(le=e.prices)==null?void 0:le.regular,locale:fe(S)})}):void 0,zoomType:ye,closeButton:ve,carouselConfig:{...s,thumbnails:s&&Object.keys(s).length?((ue=e==null?void 0:e.images)==null?void 0:ue.map(({label:t,url:i})=>{var r,c;return o(be,{title:t,alt:t,src:i,width:((r=s==null?void 0:s.thumbnailParams)==null?void 0:r.width)??200,height:((c=s==null?void 0:s.thumbnailParams)==null?void 0:c.height)??248,params:(s==null?void 0:s.thumbnailParams)??{width:200}},i)}))??[]:[]},outOfStock:!(e!=null&&e.inStock),attributes:L?void 0:o(u,{name:"Attributes",slot:n==null?void 0:n.Attributes,context:{...l},children:q!=null&&q.length?he(Fe,{children:[Ve,o("ul",{children:q})]}):null}),galleryContent:(n==null?void 0:n.GalleryContent)&&o(u,{name:"GalleryContent",slot:n.GalleryContent,context:{...l}}),infoContent:(n==null?void 0:n.InfoContent)&&o(u,{name:"InfoContent",slot:n.InfoContent,context:{...l}}),productContent:(n==null?void 0:n.Content)&&o(u,{name:"Content",slot:n.Content,context:{...l}})})})};xe.getInitialData=async function({sku:D,disableDropdownPreselection:G=!1,optionsConfig:I}){var E,L;const w=(L=(E=Pe.getConfig().models)==null?void 0:E.ProductDetails)==null?void 0:L.initialData;return w?Xe(w):W(D,{preselectFirstOption:!G,anchors:I==null?void 0:I.anchorOptions})};export{xe as ProductDetails,xe as default};
 //# sourceMappingURL=ProductDetails.js.map
