+++ scripts/__dropins__/storefront-pdp/containers/ProductDownloadableOptions.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as r}from"@dropins/tools/preact-jsx-runtime.js";import{classes as p}from"@dropins/tools/lib.js";import{s as f}from"../chunks/getProductConfigurationValues.js";import"@dropins/tools/event-bus.js";import{g as D}from"../chunks/getFetchedProductData.js";import{D as b}from"../chunks/DownloadableOptions2.js";import{u as g,g as h}from"../chunks/useProductData.js";import"@dropins/tools/preact-compat.js";import"@dropins/tools/components.js";/* empty css                           */import"@dropins/tools/i18n.js";const I=({initialData:c=null,scope:m,children:w,...o})=>{const{data:s,values:u}=g({scope:m,initialData:c,initialValues:{optionsUIDs:[]}}),d=(e,l)=>{const t=u.optionsUIDs||[];let a;l?t.includes(e)?a=t:a=[...t,e]:a=t.filter(n=>n!==e),f(n=>({...n,optionsUIDs:a}))},i=h(s,"ac_downloadable");return!s||i.length===0?null:r("div",{...o,className:p(["pdp-downloadable-options",o.className]),children:i.map((e,l)=>{var t,a;return r(b,{label:e.label,value:e.value,onLinkChange:d,currency:(a=(t=s.prices)==null?void 0:t.final)==null?void 0:a.currency,className:o.className},`downloadable-${l}`)})})};I.getInitialData=D;export{I as ProductDownloadableOptions,I as default};
+import{jsx as r}from"@dropins/tools/preact-jsx-runtime.js";import{classes as d}from"@dropins/tools/lib.js";import{setProductConfigurationValues as f,getFetchedProductData as D}from"../api.js";import"@dropins/tools/event-bus.js";import{D as b}from"../chunks/components.js";import{u as g,g as h}from"../chunks/useProductData.js";import"@dropins/tools/fetch-graphql.js";import"@dropins/tools/i18n.js";import"@dropins/tools/preact-compat.js";import"@dropins/tools/components.js";import"@dropins/tools/preact.js";import"@dropins/tools/preact-hooks.js";const I=({initialData:c=null,scope:m,children:w,...a})=>{const{data:l,values:u}=g({scope:m,initialData:c,initialValues:{optionsUIDs:[]}}),p=(o,s)=>{const t=u.optionsUIDs||[];let e;s?t.includes(o)?e=t:e=[...t,o]:e=t.filter(n=>n!==o),f(n=>({...n,optionsUIDs:e}))},i=h(l,"ac_downloadable");return!l||i.length===0?null:r("div",{...a,className:d(["pdp-downloadable-options",a.className]),children:i.map((o,s)=>{var t,e;return r(b,{label:o.label,value:o.value,onLinkChange:p,currency:(e=(t=l.prices)==null?void 0:t.final)==null?void 0:e.currency,className:a.className},`downloadable-${s}`)})})};I.getInitialData=D;export{I as ProductDownloadableOptions,I as default};
 //# sourceMappingURL=ProductDownloadableOptions.js.map
