+++ scripts/__dropins__/storefront-pdp/containers/ProductGiftCardOptions.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as p}from"@dropins/tools/preact-jsx-runtime.js";import{classes as g}from"@dropins/tools/lib.js";import{s as h}from"../chunks/getProductConfigurationValues.js";import"@dropins/tools/event-bus.js";import{s as O}from"../chunks/isProductConfigurationValid.js";import{g as C}from"../chunks/getFetchedProductData.js";import{G as V}from"../chunks/GiftCardOptions2.js";import{u as v,g as x}from"../chunks/useProductData.js";import"@dropins/tools/preact-hooks.js";import"@dropins/tools/components.js";/* empty css                       */import"@dropins/tools/i18n.js";import"@dropins/tools/preact-compat.js";const P=({initialData:l=null,scope:m,children:N,...e})=>{const{data:r,values:d}=v({scope:m,initialData:l,initialValues:{enteredOptions:[]}}),u=(i,a)=>{const n=d.enteredOptions||[],o=n.findIndex(t=>t.uid===i)!==-1?n.map(t=>t.uid===i?{...t,value:a}:t):[...n,{uid:i,value:a}];h(t=>({...t,enteredOptions:o}))},f=i=>{O(()=>i)};if(!r)return null;const c=x(r,"ac_giftcard");return c.length===0?null:p("div",{...e,className:g(["pdp-gift-card-options",e.className]),children:c.map((i,a)=>{var n,s,o;return p(V,{options:((n=i.value)==null?void 0:n.options)||[],className:e.className,optionValues:(d.enteredOptions||[]).map(t=>({id:t.uid,value:t.value})),currency:(o=(s=r.prices)==null?void 0:s.final)==null?void 0:o.currency,onOptionChange:u,onValidationChange:f},`giftcard-${a}`)})})};P.getInitialData=C;export{P as ProductGiftCardOptions,P as default};
+import{jsx as p}from"@dropins/tools/preact-jsx-runtime.js";import{classes as g}from"@dropins/tools/lib.js";import{setProductConfigurationValid as h,setProductConfigurationValues as O,getFetchedProductData as C}from"../api.js";import"@dropins/tools/event-bus.js";import{G as V}from"../chunks/components.js";import{u as v,g as x}from"../chunks/useProductData.js";import"@dropins/tools/fetch-graphql.js";import"@dropins/tools/i18n.js";import"@dropins/tools/preact-compat.js";import"@dropins/tools/components.js";import"@dropins/tools/preact.js";import"@dropins/tools/preact-hooks.js";const P=({initialData:l=null,scope:u,children:N,...e})=>{const{data:r,values:d}=v({scope:u,initialData:l,initialValues:{enteredOptions:[]}}),m=(i,a)=>{const n=d.enteredOptions||[],o=n.findIndex(t=>t.uid===i)!==-1?n.map(t=>t.uid===i?{...t,value:a}:t):[...n,{uid:i,value:a}];O(t=>({...t,enteredOptions:o}))},f=i=>{h(()=>i)};if(!r)return null;const c=x(r,"ac_giftcard");return c.length===0?null:p("div",{...e,className:g(["pdp-gift-card-options",e.className]),children:c.map((i,a)=>{var n,s,o;return p(V,{options:((n=i.value)==null?void 0:n.options)||[],className:e.className,optionValues:(d.enteredOptions||[]).map(t=>({id:t.uid,value:t.value})),currency:(o=(s=r.prices)==null?void 0:s.final)==null?void 0:o.currency,onOptionChange:m,onValidationChange:f},`giftcard-${a}`)})})};P.getInitialData=C;export{P as ProductGiftCardOptions,P as default};
 //# sourceMappingURL=ProductGiftCardOptions.js.map
