+++ scripts/__dropins__/storefront-pdp/containers/ProductHeader.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsxs as m,jsx as e,Fragment as n}from"@dropins/tools/preact-jsx-runtime.js";import{useState as f,useEffect as u}from"@dropins/tools/preact-compat.js";import{classes as _,VComponent as s}from"@dropins/tools/lib.js";import"@dropins/tools/components.js";/* empty css                  */import"@dropins/tools/preact.js";import"@dropins/tools/preact-hooks.js";/* empty css                       *//* empty css                           */import{events as v}from"@dropins/tools/event-bus.js";import{g as N}from"../chunks/getFetchedProductData.js";const g=({className:i,title:d,sku:l,hideSku:o=!1,variantName:r,variantSku:t,children:c,...a})=>m("div",{...a,className:_(["pdp-header",i]),children:[e("div",{className:"pdp-header__title",children:e(s,{node:d})}),!o&&e("div",{className:"pdp-header__sku",children:e(s,{node:l})}),(r||t)&&m("div",{className:"pdp-header__variant",children:[r&&e("span",{className:"pdp-header__variant-name",children:e(s,{node:r})}),r&&t&&e("span",{className:"pdp-header__variant-separator",children:" – "}),t&&e("span",{className:"pdp-header__variant-sku",children:e(s,{node:t})})]})]}),k=({initialData:i=null,scope:d,hideSku:l=!1,showVariantSku:o=!1,showVariantName:r=!1,children:t,...c})=>{const[a,h]=f(i);return u(()=>{const p=v.on("pdp/data",h,{scope:d});return()=>{p==null||p.off()}},[d]),a?e(g,{...c,title:e(n,{children:a.name}),sku:e(n,{children:a.sku}),hideSku:l,variantName:r&&a.variantName?e(n,{children:a.variantName}):void 0,variantSku:o&&a.variantSku?e(n,{children:a.variantSku}):void 0}):null};k.getInitialData=N;export{k as ProductHeader,k as default};
+import{jsx as r,Fragment as a}from"@dropins/tools/preact-jsx-runtime.js";import{useState as f,useEffect as s}from"@dropins/tools/preact-compat.js";import"@dropins/tools/lib.js";import"@dropins/tools/components.js";import{H as p}from"../chunks/components.js";import"@dropins/tools/preact.js";import"@dropins/tools/preact-hooks.js";import{events as c}from"@dropins/tools/event-bus.js";import{getFetchedProductData as v}from"../api.js";import"@dropins/tools/i18n.js";import"@dropins/tools/fetch-graphql.js";const h=({initialData:o=null,scope:i,hideSku:n=!1,showVariantSku:m=!1,showVariantName:d=!1,children:k,...l})=>{const[t,u]=f(o);return s(()=>{const e=c.on("pdp/data",u,{scope:i});return()=>{e==null||e.off()}},[i]),t?r(p,{...l,title:r(a,{children:t.name}),sku:r(a,{children:t.sku}),hideSku:n,variantName:d&&t.variantName?r(a,{children:t.variantName}):void 0,variantSku:m&&t.variantSku?r(a,{children:t.variantSku}):void 0}):null};h.getInitialData=v;export{h as ProductHeader,h as default};
 //# sourceMappingURL=ProductHeader.js.map
