+++ scripts/__dropins__/storefront-pdp/containers/ProductOptions.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsx as h}from"@dropins/tools/preact-jsx-runtime.js";import{useState as P,useEffect as G,useCallback as O,useMemo as H}from"@dropins/tools/preact-compat.js";import{Slot as j}from"@dropins/tools/lib.js";import{events as S}from"@dropins/tools/event-bus.js";import"@dropins/tools/components.js";/* empty css                  */import{S as J}from"../chunks/Swatches.js";import"@dropins/tools/preact.js";import"@dropins/tools/preact-hooks.js";/* empty css                       *//* empty css                           */import{c as y,s as K,f as Q,p as V}from"../chunks/fetchProductData.js";import{g as T,s as b}from"../chunks/getProductConfigurationValues.js";import{s as B}from"../chunks/isProductConfigurationValid.js";import{g as W}from"../chunks/getFetchedProductData.js";import"@dropins/tools/i18n.js";import"../fragments.js";import"@dropins/tools/fetch-graphql.js";const X=({initialData:s={data:null,optionsUIDs:[]},scope:o,hideSelectedValue:g=!1,transformer:u,slots:a,onValues:p,onErrors:m,children:Y,...M})=>{var w;const[t,F]=P(s.data),[l,L]=P(s.optionsUIDs),[N,R]=P();G(()=>{const i=S.on("pdp/data",F,{scope:o}),c=S.on("pdp/values",({optionsUIDs:f})=>{f&&L(f)},{scope:o});return()=>{i==null||i.off(),c==null||c.off()}},[]);const q=O(async i=>{var D,k,v,C;if(!(t!=null&&t.sku))return;const{persistURLParams:c,anchors:f,acdl:A}=y.getConfig()??{};let r=Object.keys(i).map(n=>i[n].value).filter(n=>n!==void 0);c&&!(t!=null&&t.isBundle)&&K({optionsUIDs:r.join(",")}),((k=(D=T({scope:o}))==null?void 0:D.optionsUIDs)==null?void 0:k.join("|"))!==(r==null?void 0:r.join("|"))&&b(n=>({...n,optionsUIDs:r}),{scope:o}),p==null||p(r);const e=await Q(t.sku,{optionsUIDs:r??[],isBundle:t==null?void 0:t.isBundle,anchors:f});S.emit("pdp/data",e,{scope:o});let U=(r==null?void 0:r.length)===((v=e==null?void 0:e.options)==null?void 0:v.length);if(f!=null&&f.length){const n=(e==null?void 0:e.optionUIDs)??[];r.every(d=>n.includes(d))||(R(e==null?void 0:e.options),b(d=>({...d,optionsUIDs:n}),{scope:o}),U=(n==null?void 0:n.length)===((C=e==null?void 0:e.options)==null?void 0:C.length))}B(()=>U,{scope:o}),A&&e&&(e==null?void 0:e.variantSku)!==t.sku&&V(e)},[t==null?void 0:t.isBundle,t==null?void 0:t.sku,p,o]),z=O(i=>{Object.keys(i).length!==0&&(B(()=>!1,{scope:o}),m==null||m(i))},[m,o]),I=H(()=>{if(t!=null&&t.options){if(!u)return t.options;try{return u(t.options)??t.options}catch(i){return console.error("ProductOptions: transformer threw an error, using original options",i),t.options}}},[t==null?void 0:t.options,u]);return!t||!I?null:h(j,{name:"Swatches",slot:a==null?void 0:a.Swatches,context:{data:t,optionsUIDs:l},children:h(J,{...M,options:I,selectedUIDs:l,defaultOptions:l,selectionsToUpdate:N,hideSelectedValue:g,disablePreselections:!((w=y.getConfig())!=null&&w.preselectFirstOption),onValues:q,onErrors:z,imageSwatchNode:a!=null&&a.SwatchImage?({imageSwatchContext:i,...c})=>h(j,{name:"SwatchImage",slot:a==null?void 0:a.SwatchImage,slotTag:"span",contentTag:"span",context:{data:t,optionsUIDs:l,defaultImageProps:c,imageSwatchContext:i}}):void 0})})};X.getInitialData=async function(s){var u;const o=await W({scope:s==null?void 0:s.scope}),g=((u=T({scope:s==null?void 0:s.scope}))==null?void 0:u.optionsUIDs)??[];return{data:o,optionsUIDs:g}};export{X as ProductOptions,X as default};
+import{jsx as I}from"@dropins/tools/preact-jsx-runtime.js";import{useState as D,useMemo as G,useEffect as H,useCallback as C}from"@dropins/tools/preact-compat.js";import{Slot as j}from"@dropins/tools/lib.js";import{events as U}from"@dropins/tools/event-bus.js";import"@dropins/tools/components.js";import{S as J}from"../chunks/components.js";import"@dropins/tools/preact.js";import"@dropins/tools/preact-hooks.js";import{config as k,s as K,c as Q,getProductConfigurationValues as y,setProductConfigurationValues as B,fetchProductData as V,i as T,setProductConfigurationValid as v,p as W,getFetchedProductData as X}from"../api.js";import{useText as Y}from"@dropins/tools/i18n.js";import"@dropins/tools/fetch-graphql.js";const Z=({initialData:n={data:null,optionsUIDs:[]},scope:i,hideSelectedValue:g=!1,transformer:p,slots:s,onValues:f,onErrors:l,children:_,...M})=>{var S;const[t,F]=D(n.data),[d,L]=D(n.optionsUIDs),[z,N]=D(),P=Y("PDP.Swatches.ChooseOption.label").label,m=G(()=>{if(t!=null&&t.options){if(!p)return t.options;try{return p(t.options)??t.options}catch(o){return console.error("ProductOptions: transformer threw an error, using original options",o),t.options}}},[t==null?void 0:t.options,p]);H(()=>{const o=U.on("pdp/data",F,{scope:i}),c=U.on("pdp/values",({optionsUIDs:a})=>{a&&L(a)},{scope:i});return()=>{o==null||o.off(),c==null||c.off()}},[]);const R=C(async o=>{var w,b;if(!(t!=null&&t.sku))return;const{persistURLParams:c,anchors:a,acdl:A}=k.getConfig()??{},E=Object.fromEntries(Object.entries(o).filter(([,u])=>u.label!==P));let r=K(E,m);c&&!(t!=null&&t.isBundle)&&Q({optionsUIDs:r.join(",")}),((b=(w=y({scope:i}))==null?void 0:w.optionsUIDs)==null?void 0:b.join("|"))!==(r==null?void 0:r.join("|"))&&B(u=>({...u,optionsUIDs:r}),{scope:i}),f==null||f(r);const e=await V(t.sku,{optionsUIDs:r??[],isBundle:t==null?void 0:t.isBundle,anchors:a});U.emit("pdp/data",e,{scope:i});let O=T(e==null?void 0:e.options,(e==null?void 0:e.optionUIDs)??r,e==null?void 0:e.isBundle);if(a!=null&&a.length){const u=(e==null?void 0:e.optionUIDs)??[];r.every(h=>u.includes(h))||(N(e==null?void 0:e.options),B(h=>({...h,optionsUIDs:u}),{scope:i}),O=T(e==null?void 0:e.options,u,e==null?void 0:e.isBundle))}v(()=>O,{scope:i}),A&&e&&(e==null?void 0:e.variantSku)!==t.sku&&W(e)},[P,t==null?void 0:t.isBundle,t==null?void 0:t.sku,f,i,m]),q=C(o=>{Object.keys(o).length!==0&&(v(()=>!1,{scope:i}),l==null||l(o))},[l,i]);return!t||!m?null:I(j,{name:"Swatches",slot:s==null?void 0:s.Swatches,context:{data:t,optionsUIDs:d},children:I(J,{...M,options:m,selectedUIDs:d,defaultOptions:d,selectionsToUpdate:z,hideSelectedValue:g,disablePreselections:!((S=k.getConfig())!=null&&S.preselectFirstOption),onValues:R,onErrors:q,imageSwatchNode:s!=null&&s.SwatchImage?({imageSwatchContext:o,...c})=>I(j,{name:"SwatchImage",slot:s==null?void 0:s.SwatchImage,slotTag:"span",contentTag:"span",context:{data:t,optionsUIDs:d,defaultImageProps:c,imageSwatchContext:o}}):void 0})})};Z.getInitialData=async function(n){var p;const i=await X({scope:n==null?void 0:n.scope}),g=((p=y({scope:n==null?void 0:n.scope}))==null?void 0:p.optionsUIDs)??[];return{data:i,optionsUIDs:g}};export{Z as ProductOptions,Z as default};
 //# sourceMappingURL=ProductOptions.js.map
