+++ scripts/__dropins__/storefront-pdp/containers/ProductPrice.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsxs as j,jsx as t}from"@dropins/tools/preact-jsx-runtime.js";import{useState as T,useEffect as C}from"@dropins/tools/preact-compat.js";import{classes as g,VComponent as x,toLanguageTag as v}from"@dropins/tools/lib.js";import{Price as A}from"@dropins/tools/components.js";/* empty css                  */import{P as q}from"../chunks/PriceRange.js";import"@dropins/tools/preact.js";import{useText as b}from"@dropins/tools/i18n.js";import"@dropins/tools/preact-hooks.js";/* empty css                       *//* empty css                           */import{events as R}from"@dropins/tools/event-bus.js";import{g as B}from"../chunks/getFetchedProductData.js";import{c as F}from"../chunks/fetchProductData.js";import"../fragments.js";import"@dropins/tools/fetch-graphql.js";import"../chunks/getProductConfigurationValues.js";import"../chunks/isProductConfigurationValid.js";const I=({className:p,price:i,specialPrice:c,tiers:a,children:s,...e})=>{const n=b("PDP.Product.RegularPrice.label").label,r=b("PDP.Product.SpecialPrice.label").label,o=b("PDP.Product.TierPrice.label").label;return j("div",{...e,className:g(["pdp-price",p]),children:[c&&t(x,{node:c,className:g(["pdp-price__amount-special","pdp-price__amount"]),"aria-label":r,role:"text"}),i&&t(x,{node:i,className:g(["pdp-price__amount-regular","pdp-price__amount--grey"]),"aria-label":n,role:"text"}),a&&t(x,{node:a,className:g(["pdp-price__tiers"]),"aria-label":o,role:"text"})]})},M=({locale:p,tiers:i,regularPrice:c,...a})=>{const s=b("PDP.Product.TierPricing.label").label;return!i||i.length===0?null:t("div",{className:"pdp-price-tiers",...a,children:i.map((e,n)=>{var m,u;const r=((m=e==null?void 0:e.tier)==null?void 0:m.amount)??0,o=(e==null?void 0:e.quantity)??0,d=((u=e==null?void 0:e.tier)==null?void 0:u.currency)??"";let l=0;c&&r&&r<c&&(l=Math.round((c-r)/c*100));const P=s.replace("{quantity}",o.toString()).split("{price}")[0],f=s.replace("{quantity}",o.toString()).split("{price}")[1].replace("{percent}",l>0?l.toString():"0");return t("div",{className:"pdp-price-tiers__tier",children:j("span",{className:"pdp-price-tiers__text",children:[P,t(A,{amount:r,currency:d,locale:p,className:"pdp-price-tiers__price"}),f]})},n)})})},U=({initialData:p=null,scope:i,children:c,...a})=>{var d,l,P,f,m,u,N,y,D,L,S;const{globalLocale:s="en-US"}=F.getConfig(),[e,n]=T(p),[r,o]=T(s);return C(()=>{const _=R.on("pdp/data",n,{scope:i}),h=R.on("locale",o);return()=>{_==null||_.off(),h==null||h.off()}},[i]),!e||!e.prices?null:t(I,{...a,specialPrice:(d=e.prices)!=null&&d.visible?t(q,{...(l=e.prices)==null?void 0:l.final,locale:v(r)}):void 0,price:(P=e.prices)!=null&&P.visible&&((f=e.prices)!=null&&f.regular)&&((m=e.prices)==null?void 0:m.final.amount)!==((u=e.prices)==null?void 0:u.regular.amount)?t(q,{...(N=e.prices)==null?void 0:N.regular,locale:v(r)}):void 0,tiers:(y=e.prices)!=null&&y.visible&&((L=(D=e.prices)==null?void 0:D.tiers)==null?void 0:L.length)>0?t(M,{tiers:e.prices.tiers,locale:v(r),regularPrice:(S=e.prices.regular)==null?void 0:S.amount}):void 0})};U.getInitialData=B;export{U as ProductPrice,U as default};
+import{jsx as e}from"@dropins/tools/preact-jsx-runtime.js";import{useState as v,useEffect as C}from"@dropins/tools/preact-compat.js";import{toLanguageTag as a}from"@dropins/tools/lib.js";import"@dropins/tools/components.js";import{P as F,b as I,a as h}from"../chunks/components.js";import"@dropins/tools/preact.js";import"@dropins/tools/preact-hooks.js";import{events as D}from"@dropins/tools/event-bus.js";import{config as R,getFetchedProductData as U}from"../api.js";import"@dropins/tools/i18n.js";import"@dropins/tools/fetch-graphql.js";const k=({initialData:L=null,scope:c,children:q,...x})=>{var s,l,p,n,m,u,f,g,d,P,b;const{globalLocale:S="en-US"}=R.getConfig(),[r,T]=v(L),[i,j]=v(S);return C(()=>{const t=D.on("pdp/data",T,{scope:c}),o=D.on("locale",j);return()=>{t==null||t.off(),o==null||o.off()}},[c]),!r||!r.prices?null:e(F,{...x,specialPrice:(s=r.prices)!=null&&s.visible?e(h,{...(l=r.prices)==null?void 0:l.final,locale:a(i)}):void 0,price:(p=r.prices)!=null&&p.visible&&((n=r.prices)!=null&&n.regular)&&((m=r.prices)==null?void 0:m.final.amount)!==((u=r.prices)==null?void 0:u.regular.amount)?e(h,{...(f=r.prices)==null?void 0:f.regular,locale:a(i)}):void 0,tiers:(g=r.prices)!=null&&g.visible&&((P=(d=r.prices)==null?void 0:d.tiers)==null?void 0:P.length)>0?e(I,{tiers:r.prices.tiers,locale:a(i),regularPrice:(b=r.prices.regular)==null?void 0:b.amount}):void 0})};k.getInitialData=U;export{k as ProductPrice,k as default};
 //# sourceMappingURL=ProductPrice.js.map
